Move all to src folder
