+++ jsconfig.json
@@ -3,13 +3,4 @@
 		"checkJs": true,
 		"target": "esnext",
 	},
-	"include": [
-		"public/**/*",
-		"tests/**/*",
-		"src/**/*"
-	],
-	"exclude": [
-		"public/web_modules/",
-		"node_modules",
-	]
 }
