--- public/web_modules/gsap.js
@@ -1,21 +1,0 @@
-function t(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function e(t,e){t.prototype=Object.create(e.prototype),t.prototype.constructor=t,t.__proto__=e}
-/*!
- * GSAP 3.11.0
- * https://greensock.com
- *
- * @license Copyright 2008-2022, GreenSock. All rights reserved.
- * Subject to the terms at https://greensock.com/standard-license or for
- * Club GreenSock members, the agreement issued with that membership.
- * @author: Jack Doyle, jack@greensock.com
-*/var r,i,n,s,a,o,u,h,f,l,c,p,_,d,m,g={autoSleep:120,force3D:"auto",nullTargetWarn:1,units:{lineHeight:""}},v={duration:.5,overwrite:!1,delay:0},y=2*Math.PI,x=y/4,T=0,w=Math.sqrt,b=Math.cos,M=Math.sin,O=function(t){return"string"==typeof t},k=function(t){return"function"==typeof t},C=function(t){return"number"==typeof t},A=function(t){return void 0===t},D=function(t){return"object"==typeof t},P=function(t){return!1!==t},S=function(){return"undefined"!=typeof window},E=function(t){return k(t)||O(t)},R="function"==typeof ArrayBuffer&&ArrayBuffer.isView||function(){},z=Array.isArray,F=/(?:-?\.?\d|\.)+/gi,B=/[-+=.]*\d+[.e\-+]*\d*[e\-+]*\d*/g,L=/[-+=.]*\d+[.e-]*\d*[a-z%]*/g,I=/[-+=.]*\d+\.?\d*(?:e-|e\+)?\d*/gi,Y=/[+-]=-?[.\d]+/,U=/[^,'"\[\]\s]+/gi,X=/^[+\-=e\s\d]*\d+[.\d]*([a-z]*|%)\s*$/i,N={},q={},V=function(t){return(q=vt(t,N))&&dr},j=function(t,e){return console.warn("Invalid property",t,"set to",e,"Missing plugin? gsap.registerPlugin()")},G=function(t,e){return!e&&console.warn(t)},W=function(t,e){return t&&(N[t]=e)&&q&&(q[t]=e)||N},Q=function(){return 0},H={suppressEvents:!0,isStart:!0},Z={suppressEvents:!0},$={},J=[],K={},tt={},et={},rt=30,it=[],nt="",st=function(t){var e,r,i=t[0];if(D(i)||k(i)||(t=[t]),!(e=(i._gsap||{}).harness)){for(r=it.length;r--&&!it[r].targetTest(i););e=it[r]}for(r=t.length;r--;)t[r]&&(t[r]._gsap||(t[r]._gsap=new Pe(t[r],e)))||t.splice(r,1);return t},at=function(t){return t._gsap||st(Ht(t))[0]._gsap},ot=function(t,e,r){return(r=t[e])&&k(r)?t[e]():A(r)&&t.getAttribute&&t.getAttribute(e)||r},ut=function(t,e){return(t=t.split(",")).forEach(e)||t},ht=function(t){return Math.round(1e5*t)/1e5||0},ft=function(t){return Math.round(1e7*t)/1e7||0},lt=function(t,e){var r=e.charAt(0),i=parseFloat(e.substr(2));return t=parseFloat(t),"+"===r?t+i:"-"===r?t-i:"*"===r?t*i:t/i},ct=function(t,e){for(var r=e.length,i=0;t.indexOf(e[i])<0&&++i<r;);return i<r},pt=function(){var t,e,r=J.length,i=J.slice(0);for(K={},J.length=0,t=0;t<r;t++)(e=i[t])&&e._lazy&&(e.render(e._lazy[0],e._lazy[1],!0)._lazy=0)},_t=function(t,e,r,n){J.length&&pt(),t.render(e,r,n||i),J.length&&pt()},dt=function(t){var e=parseFloat(t);return(e||0===e)&&(t+"").match(U).length<2?e:O(t)?t.trim():t},mt=function(t){return t},gt=function(t,e){for(var r in e)r in t||(t[r]=e[r]);return t},vt=function(t,e){for(var r in e)t[r]=e[r];return t},yt=function t(e,r){for(var i in r)"__proto__"!==i&&"constructor"!==i&&"prototype"!==i&&(e[i]=D(r[i])?t(e[i]||(e[i]={}),r[i]):r[i]);return e},xt=function(t,e){var r,i={};for(r in t)r in e||(i[r]=t[r]);return i},Tt=function(t){var e,r=t.parent||s,i=t.keyframes?(e=z(t.keyframes),function(t,r){for(var i in r)i in t||"duration"===i&&e||"ease"===i||(t[i]=r[i])}):gt;if(P(t.inherit))for(;r;)i(t,r.vars.defaults),r=r.parent||r._dp;return t},wt=function(t,e,r,i,n){void 0===r&&(r="_first"),void 0===i&&(i="_last");var s,a=t[i];if(n)for(s=e[n];a&&a[n]>s;)a=a._prev;return a?(e._next=a._next,a._next=e):(e._next=t[r],t[r]=e),e._next?e._next._prev=e:t[i]=e,e._prev=a,e.parent=e._dp=t,e},bt=function(t,e,r,i){void 0===r&&(r="_first"),void 0===i&&(i="_last");var n=e._prev,s=e._next;n?n._next=s:t[r]===e&&(t[r]=s),s?s._prev=n:t[i]===e&&(t[i]=n),e._next=e._prev=e.parent=null},Mt=function(t,e){t.parent&&(!e||t.parent.autoRemoveChildren)&&t.parent.remove(t),t._act=0},Ot=function(t,e){if(t&&(!e||e._end>t._dur||e._start<0))for(var r=t;r;)r._dirty=1,r=r.parent;return t},kt=function(t){for(var e=t.parent;e&&e.parent;)e._dirty=1,e.totalDuration(),e=e.parent;return t},Ct=function(t,e,r,n){return t._startAt&&(i?t._startAt.revert(Z):t.vars.immediateRender&&!t.vars.autoRevert||t._startAt.render(e,!0,n))},At=function(t){return t._repeat?Dt(t._tTime,t=t.duration()+t._rDelay)*t:0},Dt=function(t,e){var r=Math.floor(t/=e);return t&&r===t?r-1:r},Pt=function(t,e){return(t-e._start)*e._ts+(e._ts>=0?0:e._dirty?e.totalDuration():e._tDur)},St=function(t){return t._end=ft(t._start+(t._tDur/Math.abs(t._ts||t._rts||1e-8)||0))},Et=function(t,e){var r=t._dp;return r&&r.smoothChildTiming&&t._ts&&(t._start=ft(r._time-(t._ts>0?e/t._ts:((t._dirty?t.totalDuration():t._tDur)-e)/-t._ts)),St(t),r._dirty||Ot(r,t)),t},Rt=function(t,e){var r;if((e._time||e._initted&&!e._dur)&&(r=Pt(t.rawTime(),e),(!e._dur||Vt(0,e.totalDuration(),r)-e._tTime>1e-8)&&e.render(r,!0)),Ot(t,e)._dp&&t._initted&&t._time>=t._dur&&t._ts){if(t._dur<t.duration())for(r=t;r._dp;)r.rawTime()>=0&&r.totalTime(r._tTime),r=r._dp;t._zTime=-1e-8}},zt=function(t,e,r,i){return e.parent&&Mt(e),e._start=ft((C(r)?r:r||t!==s?Xt(t,r,e):t._time)+e._delay),e._end=ft(e._start+(e.totalDuration()/Math.abs(e.timeScale())||0)),wt(t,e,"_first","_last",t._sort?"_start":0),Lt(e)||(t._recent=e),i||Rt(t,e),t._ts<0&&Et(t,t._tTime),t},Ft=function(t,e){return(N.ScrollTrigger||j("scrollTrigger",e))&&N.ScrollTrigger.create(e,t)},Bt=function(t,e,r,i){return Ie(t,e),t._initted?!r&&t._pt&&(t._dur&&!1!==t.vars.lazy||!t._dur&&t.vars.lazy)&&f!==ge.frame?(J.push(t),t._lazy=[e,i],1):void 0:1},Lt=function(t){var e=t.data;return"isFromStart"===e||"isStart"===e},It=function(t,e,r,i){var n=t._repeat,s=ft(e)||0,a=t._tTime/t._tDur;return a&&!i&&(t._time*=s/t._dur),t._dur=s,t._tDur=n?n<0?1e10:ft(s*(n+1)+t._rDelay*n):s,a>0&&!i?Et(t,t._tTime=t._tDur*a):t.parent&&St(t),r||Ot(t.parent,t),t},Yt=function(t){return t instanceof Ee?Ot(t):It(t,t._dur)},Ut={_start:0,endTime:Q,totalDuration:Q},Xt=function t(e,r,i){var n,s,a,o=e.labels,u=e._recent||Ut,h=e.duration()>=1e8?u.endTime(!1):e._dur;return O(r)&&(isNaN(r)||r in o)?(s=r.charAt(0),a="%"===r.substr(-1),n=r.indexOf("="),"<"===s||">"===s?(n>=0&&(r=r.replace(/=/,"")),("<"===s?u._start:u.endTime(u._repeat>=0))+(parseFloat(r.substr(1))||0)*(a?(n<0?u:i).totalDuration()/100:1)):n<0?(r in o||(o[r]=h),o[r]):(s=parseFloat(r.charAt(n-1)+r.substr(n+1)),a&&i&&(s=s/100*(z(i)?i[0]:i).totalDuration()),n>1?t(e,r.substr(0,n-1),i)+s:h+s)):null==r?h:+r},Nt=function(t,e,r){var i,n,s=C(e[1]),a=(s?2:1)+(t<2?0:1),o=e[a];if(s&&(o.duration=e[1]),o.parent=r,t){for(i=o,n=r;n&&!("immediateRender"in i);)i=n.vars.defaults||{},n=P(n.vars.inherit)&&n.parent;o.immediateRender=P(i.immediateRender),t<2?o.runBackwards=1:o.startAt=e[a-1]}return new qe(e[0],o,e[a+1])},qt=function(t,e){return t||0===t?e(t):e},Vt=function(t,e,r){return r<t?t:r>e?e:r},jt=function(t,e){return O(t)&&(e=X.exec(t))?e[1]:""},Gt=[].slice,Wt=function(t,e){return t&&D(t)&&"length"in t&&(!e&&!t.length||t.length-1 in t&&D(t[0]))&&!t.nodeType&&t!==a},Qt=function(t,e,r){return void 0===r&&(r=[]),t.forEach((function(t){var i;return O(t)&&!e||Wt(t,1)?(i=r).push.apply(i,Ht(t)):r.push(t)}))||r},Ht=function(t,e,r){return n&&!e&&n.selector?n.selector(t):!O(t)||r||!o&&ve()?z(t)?Qt(t,r):Wt(t)?Gt.call(t,0):t?[t]:[]:Gt.call((e||u).querySelectorAll(t),0)},Zt=function(t){return t=Ht(t)[0]||G("Invalid scope")||{},function(e){var r=t.current||t.nativeElement||t;return Ht(e,r.querySelectorAll?r:r===t?G("Invalid scope")||u.createElement("div"):t)}},$t=function(t){return t.sort((function(){return.5-Math.random()}))},Jt=function(t){if(k(t))return t;var e=D(t)?t:{each:t},r=Oe(e.ease),i=e.from||0,n=parseFloat(e.base)||0,s={},a=i>0&&i<1,o=isNaN(i)||a,u=e.axis,h=i,f=i;return O(i)?h=f={center:.5,edges:.5,end:1}[i]||0:!a&&o&&(h=i[0],f=i[1]),function(t,a,l){var c,p,_,d,m,g,v,y,x,T=(l||e).length,b=s[T];if(!b){if(!(x="auto"===e.grid?0:(e.grid||[1,1e8])[1])){for(v=-1e8;v<(v=l[x++].getBoundingClientRect().left)&&x<T;);x--}for(b=s[T]=[],c=o?Math.min(x,T)*h-.5:i%x,p=1e8===x?0:o?T*f/x-.5:i/x|0,v=0,y=1e8,g=0;g<T;g++)_=g%x-c,d=p-(g/x|0),b[g]=m=u?Math.abs("y"===u?d:_):w(_*_+d*d),m>v&&(v=m),m<y&&(y=m);"random"===i&&$t(b),b.max=v-y,b.min=y,b.v=T=(parseFloat(e.amount)||parseFloat(e.each)*(x>T?T-1:u?"y"===u?T/x:x:Math.max(x,T/x))||0)*("edges"===i?-1:1),b.b=T<0?n-T:n,b.u=jt(e.amount||e.each)||0,r=r&&T<0?be(r):r}return T=(b[t]-b.min)/b.max||0,ft(b.b+(r?r(T):T)*b.v)+b.u}},Kt=function(t){var e=Math.pow(10,((t+"").split(".")[1]||"").length);return function(r){var i=ft(Math.round(parseFloat(r)/t)*t*e);return(i-i%1)/e+(C(r)?0:jt(r))}},te=function(t,e){var r,i,n=z(t);return!n&&D(t)&&(r=n=t.radius||1e8,t.values?(t=Ht(t.values),(i=!C(t[0]))&&(r*=r)):t=Kt(t.increment)),qt(e,n?k(t)?function(e){return i=t(e),Math.abs(i-e)<=r?i:e}:function(e){for(var n,s,a=parseFloat(i?e.x:e),o=parseFloat(i?e.y:0),u=1e8,h=0,f=t.length;f--;)(n=i?(n=t[f].x-a)*n+(s=t[f].y-o)*s:Math.abs(t[f]-a))<u&&(u=n,h=f);return h=!r||u<=r?t[h]:e,i||h===e||C(e)?h:h+jt(e)}:Kt(t))},ee=function(t,e,r,i){return qt(z(t)?!e:!0===r?!!(r=0):!i,(function(){return z(t)?t[~~(Math.random()*t.length)]:(r=r||1e-5)&&(i=r<1?Math.pow(10,(r+"").length-2):1)&&Math.floor(Math.round((t-r/2+Math.random()*(e-t+.99*r))/r)*r*i)/i}))},re=function(t,e,r){return qt(r,(function(r){return t[~~e(r)]}))},ie=function(t){for(var e,r,i,n,s=0,a="";~(e=t.indexOf("random(",s));)i=t.indexOf(")",e),n="["===t.charAt(e+7),r=t.substr(e+7,i-e-7).match(n?U:F),a+=t.substr(s,e-s)+ee(n?r:+r[0],n?0:+r[1],+r[2]||1e-5),s=i+1;return a+t.substr(s,t.length-s)},ne=function(t,e,r,i,n){var s=e-t,a=i-r;return qt(n,(function(e){return r+((e-t)/s*a||0)}))},se=function(t,e,r){var i,n,s,a=t.labels,o=1e8;for(i in a)(n=a[i]-e)<0==!!r&&n&&o>(n=Math.abs(n))&&(s=i,o=n);return s},ae=function(t,e,r){var i,s,a,o=t.vars,u=o[e],h=n,f=t._ctx;if(u)return i=o[e+"Params"],s=o.callbackScope||t,r&&J.length&&pt(),f&&(n=f),a=i?u.apply(s,i):u.call(s),n=h,a},oe=function(t){return Mt(t),t.scrollTrigger&&t.scrollTrigger.kill(!1),t.progress()<1&&ae(t,"onInterrupt"),t},ue=function(t){var e=(t=!t.name&&t.default||t).name,r=k(t),i=e&&!r&&t.init?function(){this._props=[]}:t,n={init:Q,render:Je,add:Be,kill:tr,modifier:Ke,rawVars:0},s={targetTest:0,get:0,getSetter:Qe,aliases:{},register:0};if(ve(),t!==i){if(tt[e])return;gt(i,gt(xt(t,n),s)),vt(i.prototype,vt(n,xt(t,s))),tt[i.prop=e]=i,t.targetTest&&(it.push(i),$[e]=1),e=("css"===e?"CSS":e.charAt(0).toUpperCase()+e.substr(1))+"Plugin"}W(e,i),t.register&&t.register(dr,i,ir)},he={aqua:[0,255,255],lime:[0,255,0],silver:[192,192,192],black:[0,0,0],maroon:[128,0,0],teal:[0,128,128],blue:[0,0,255],navy:[0,0,128],white:[255,255,255],olive:[128,128,0],yellow:[255,255,0],orange:[255,165,0],gray:[128,128,128],purple:[128,0,128],green:[0,128,0],red:[255,0,0],pink:[255,192,203],cyan:[0,255,255],transparent:[255,255,255,0]},fe=function(t,e,r){return 255*(6*(t+=t<0?1:t>1?-1:0)<1?e+(r-e)*t*6:t<.5?r:3*t<2?e+(r-e)*(2/3-t)*6:e)+.5|0},le=function(t,e,r){var i,n,s,a,o,u,h,f,l,c,p=t?C(t)?[t>>16,t>>8&255,255&t]:0:he.black;if(!p){if(","===t.substr(-1)&&(t=t.substr(0,t.length-1)),he[t])p=he[t];else if("#"===t.charAt(0)){if(t.length<6&&(i=t.charAt(1),n=t.charAt(2),s=t.charAt(3),t="#"+i+i+n+n+s+s+(5===t.length?t.charAt(4)+t.charAt(4):"")),9===t.length)return[(p=parseInt(t.substr(1,6),16))>>16,p>>8&255,255&p,parseInt(t.substr(7),16)/255];p=[(t=parseInt(t.substr(1),16))>>16,t>>8&255,255&t]}else if("hsl"===t.substr(0,3))if(p=c=t.match(F),e){if(~t.indexOf("="))return p=t.match(B),r&&p.length<4&&(p[3]=1),p}else a=+p[0]%360/360,o=+p[1]/100,i=2*(u=+p[2]/100)-(n=u<=.5?u*(o+1):u+o-u*o),p.length>3&&(p[3]*=1),p[0]=fe(a+1/3,i,n),p[1]=fe(a,i,n),p[2]=fe(a-1/3,i,n);else p=t.match(F)||he.transparent;p=p.map(Number)}return e&&!c&&(i=p[0]/255,n=p[1]/255,s=p[2]/255,u=((h=Math.max(i,n,s))+(f=Math.min(i,n,s)))/2,h===f?a=o=0:(l=h-f,o=u>.5?l/(2-h-f):l/(h+f),a=h===i?(n-s)/l+(n<s?6:0):h===n?(s-i)/l+2:(i-n)/l+4,a*=60),p[0]=~~(a+.5),p[1]=~~(100*o+.5),p[2]=~~(100*u+.5)),r&&p.length<4&&(p[3]=1),p},ce=function(t){var e=[],r=[],i=-1;return t.split(_e).forEach((function(t){var n=t.match(L)||[];e.push.apply(e,n),r.push(i+=n.length+1)})),e.c=r,e},pe=function(t,e,r){var i,n,s,a,o="",u=(t+o).match(_e),h=e?"hsla(":"rgba(",f=0;if(!u)return t;if(u=u.map((function(t){return(t=le(t,e,1))&&h+(e?t[0]+","+t[1]+"%,"+t[2]+"%,"+t[3]:t.join(","))+")"})),r&&(s=ce(t),(i=r.c).join(o)!==s.c.join(o)))for(a=(n=t.replace(_e,"1").split(L)).length-1;f<a;f++)o+=n[f]+(~i.indexOf(f)?u.shift()||h+"0,0,0,0)":(s.length?s:u.length?u:r).shift());if(!n)for(a=(n=t.split(_e)).length-1;f<a;f++)o+=n[f]+u[f];return o+n[a]},_e=function(){var t,e="(?:\\b(?:(?:rgb|rgba|hsl|hsla)\\(.+?\\))|\\B#(?:[0-9a-f]{3,4}){1,2}\\b";for(t in he)e+="|"+t+"\\b";return new RegExp(e+")","gi")}(),de=/hsl[a]?\(/,me=function(t){var e,r=t.join(" ");if(_e.lastIndex=0,_e.test(r))return e=de.test(r),t[1]=pe(t[1],e),t[0]=pe(t[0],e,ce(t[1])),!0},ge=function(){var t,e,r,i,n,s,f=Date.now,l=500,p=33,_=f(),d=_,m=1e3/240,g=m,v=[],y=function r(a){var o,u,h,c,y=f()-d,x=!0===a;if(y>l&&(_+=y-p),((o=(h=(d+=y)-_)-g)>0||x)&&(c=++i.frame,n=h-1e3*i.time,i.time=h/=1e3,g+=o+(o>=m?4:m-o),u=1),x||(t=e(r)),u)for(s=0;s<v.length;s++)v[s](h,n,c,a)};return i={time:0,frame:0,tick:function(){y(!0)},deltaRatio:function(t){return n/(1e3/(t||60))},wake:function(){h&&(!o&&S()&&(a=o=window,u=a.document||{},N.gsap=dr,(a.gsapVersions||(a.gsapVersions=[])).push(dr.version),V(q||a.GreenSockGlobals||!a.gsap&&a||{}),r=a.requestAnimationFrame),t&&i.sleep(),e=r||function(t){return setTimeout(t,g-1e3*i.time+1|0)},c=1,y(2))},sleep:function(){(r?a.cancelAnimationFrame:clearTimeout)(t),c=0,e=Q},lagSmoothing:function(t,e){l=t||1/1e-8,p=Math.min(e,l,0)},fps:function(t){m=1e3/(t||240),g=1e3*i.time+m},add:function(t,e,r){var n=e?function(e,r,s,a){t(e,r,s,a),i.remove(n)}:t;return i.remove(t),v[r?"unshift":"push"](n),ve(),n},remove:function(t,e){~(e=v.indexOf(t))&&v.splice(e,1)&&s>=e&&s--},_listeners:v}}(),ve=function(){return!c&&ge.wake()},ye={},xe=/^[\d.\-M][\d.\-,\s]/,Te=/["']/g,we=function(t){for(var e,r,i,n={},s=t.substr(1,t.length-3).split(":"),a=s[0],o=1,u=s.length;o<u;o++)r=s[o],e=o!==u-1?r.lastIndexOf(","):r.length,i=r.substr(0,e),n[a]=isNaN(i)?i.replace(Te,"").trim():+i,a=r.substr(e+1).trim();return n},be=function(t){return function(e){return 1-t(1-e)}},Me=function t(e,r){for(var i,n=e._first;n;)n instanceof Ee?t(n,r):!n.vars.yoyoEase||n._yoyo&&n._repeat||n._yoyo===r||(n.timeline?t(n.timeline,r):(i=n._ease,n._ease=n._yEase,n._yEase=i,n._yoyo=r)),n=n._next},Oe=function(t,e){return t&&(k(t)?t:ye[t]||function(t){var e,r,i,n,s=(t+"").split("("),a=ye[s[0]];return a&&s.length>1&&a.config?a.config.apply(null,~t.indexOf("{")?[we(s[1])]:(e=t,r=e.indexOf("(")+1,i=e.indexOf(")"),n=e.indexOf("(",r),e.substring(r,~n&&n<i?e.indexOf(")",i+1):i)).split(",").map(dt)):ye._CE&&xe.test(t)?ye._CE("",t):a}(t))||e},ke=function(t,e,r,i){void 0===r&&(r=function(t){return 1-e(1-t)}),void 0===i&&(i=function(t){return t<.5?e(2*t)/2:1-e(2*(1-t))/2});var n,s={easeIn:e,easeOut:r,easeInOut:i};return ut(t,(function(t){for(var e in ye[t]=N[t]=s,ye[n=t.toLowerCase()]=r,s)ye[n+("easeIn"===e?".in":"easeOut"===e?".out":".inOut")]=ye[t+"."+e]=s[e]})),s},Ce=function(t){return function(e){return e<.5?(1-t(1-2*e))/2:.5+t(2*(e-.5))/2}},Ae=function t(e,r,i){var n=r>=1?r:1,s=(i||(e?.3:.45))/(r<1?r:1),a=s/y*(Math.asin(1/n)||0),o=function(t){return 1===t?1:n*Math.pow(2,-10*t)*M((t-a)*s)+1},u="out"===e?o:"in"===e?function(t){return 1-o(1-t)}:Ce(o);return s=y/s,u.config=function(r,i){return t(e,r,i)},u},De=function t(e,r){void 0===r&&(r=1.70158);var i=function(t){return t?--t*t*((r+1)*t+r)+1:0},n="out"===e?i:"in"===e?function(t){return 1-i(1-t)}:Ce(i);return n.config=function(r){return t(e,r)},n};ut("Linear,Quad,Cubic,Quart,Quint,Strong",(function(t,e){var r=e<5?e+1:e;ke(t+",Power"+(r-1),e?function(t){return Math.pow(t,r)}:function(t){return t},(function(t){return 1-Math.pow(1-t,r)}),(function(t){return t<.5?Math.pow(2*t,r)/2:1-Math.pow(2*(1-t),r)/2}))})),ye.Linear.easeNone=ye.none=ye.Linear.easeIn,ke("Elastic",Ae("in"),Ae("out"),Ae()),p=7.5625,d=1/(_=2.75),ke("Bounce",(function(t){return 1-m(1-t)}),m=function(t){return t<d?p*t*t:t<.7272727272727273?p*Math.pow(t-1.5/_,2)+.75:t<.9090909090909092?p*(t-=2.25/_)*t+.9375:p*Math.pow(t-2.625/_,2)+.984375}),ke("Expo",(function(t){return t?Math.pow(2,10*(t-1)):0})),ke("Circ",(function(t){return-(w(1-t*t)-1)})),ke("Sine",(function(t){return 1===t?1:1-b(t*x)})),ke("Back",De("in"),De("out"),De()),ye.SteppedEase=ye.steps=N.SteppedEase={config:function(t,e){void 0===t&&(t=1);var r=1/t,i=t+(e?0:1),n=e?1:0;return function(t){return((i*Vt(0,1-1e-8,t)|0)+n)*r}}},v.ease=ye["quad.out"],ut("onComplete,onUpdate,onStart,onRepeat,onReverseComplete,onInterrupt",(function(t){return nt+=t+","+t+"Params,"}));var Pe=function(t,e){this.id=T++,t._gsap=this,this.target=t,this.harness=e,this.get=e?e.get:ot,this.set=e?e.getSetter:Qe},Se=function(){function t(t){this.vars=t,this._delay=+t.delay||0,(this._repeat=t.repeat===1/0?-2:t.repeat||0)&&(this._rDelay=t.repeatDelay||0,this._yoyo=!!t.yoyo||!!t.yoyoEase),this._ts=1,It(this,+t.duration,1,1),this.data=t.data,n&&(this._ctx=n,n.data.push(this)),c||ge.wake()}var e=t.prototype;return e.delay=function(t){return t||0===t?(this.parent&&this.parent.smoothChildTiming&&this.startTime(this._start+t-this._delay),this._delay=t,this):this._delay},e.duration=function(t){return arguments.length?this.totalDuration(this._repeat>0?t+(t+this._rDelay)*this._repeat:t):this.totalDuration()&&this._dur},e.totalDuration=function(t){return arguments.length?(this._dirty=0,It(this,this._repeat<0?t:(t-this._repeat*this._rDelay)/(this._repeat+1))):this._tDur},e.totalTime=function(t,e){if(ve(),!arguments.length)return this._tTime;var r=this._dp;if(r&&r.smoothChildTiming&&this._ts){for(Et(this,t),!r._dp||r.parent||Rt(r,this);r&&r.parent;)r.parent._time!==r._start+(r._ts>=0?r._tTime/r._ts:(r.totalDuration()-r._tTime)/-r._ts)&&r.totalTime(r._tTime,!0),r=r.parent;!this.parent&&this._dp.autoRemoveChildren&&(this._ts>0&&t<this._tDur||this._ts<0&&t>0||!this._tDur&&!t)&&zt(this._dp,this,this._start-this._delay)}return(this._tTime!==t||!this._dur&&!e||this._initted&&1e-8===Math.abs(this._zTime)||!t&&!this._initted&&(this.add||this._ptLookup))&&(this._ts||(this._pTime=t),_t(this,t,e)),this},e.time=function(t,e){return arguments.length?this.totalTime(Math.min(this.totalDuration(),t+At(this))%(this._dur+this._rDelay)||(t?this._dur:0),e):this._time},e.totalProgress=function(t,e){return arguments.length?this.totalTime(this.totalDuration()*t,e):this.totalDuration()?Math.min(1,this._tTime/this._tDur):this.ratio},e.progress=function(t,e){return arguments.length?this.totalTime(this.duration()*(!this._yoyo||1&this.iteration()?t:1-t)+At(this),e):this.duration()?Math.min(1,this._time/this._dur):this.ratio},e.iteration=function(t,e){var r=this.duration()+this._rDelay;return arguments.length?this.totalTime(this._time+(t-1)*r,e):this._repeat?Dt(this._tTime,r)+1:1},e.timeScale=function(t){if(!arguments.length)return-1e-8===this._rts?0:this._rts;if(this._rts===t)return this;var e=this.parent&&this._ts?Pt(this.parent._time,this):this._tTime;return this._rts=+t||0,this._ts=this._ps||-1e-8===t?0:this._rts,this.totalTime(Vt(-this._delay,this._tDur,e),!0),St(this),kt(this)},e.paused=function(t){return arguments.length?(this._ps!==t&&(this._ps=t,t?(this._pTime=this._tTime||Math.max(-this._delay,this.rawTime()),this._ts=this._act=0):(ve(),this._ts=this._rts,this.totalTime(this.parent&&!this.parent.smoothChildTiming?this.rawTime():this._tTime||this._pTime,1===this.progress()&&1e-8!==Math.abs(this._zTime)&&(this._tTime-=1e-8)))),this):this._ps},e.startTime=function(t){if(arguments.length){this._start=t;var e=this.parent||this._dp;return e&&(e._sort||!this.parent)&&zt(e,this,t-this._delay),this}return this._start},e.endTime=function(t){return this._start+(P(t)?this.totalDuration():this.duration())/Math.abs(this._ts||1)},e.rawTime=function(t){var e=this.parent||this._dp;return e?t&&(!this._ts||this._repeat&&this._time&&this.totalProgress()<1)?this._tTime%(this._dur+this._rDelay):this._ts?Pt(e.rawTime(t),this):this._tTime:this._tTime},e.revert=function(t){void 0===t&&(t=Z);var e=i;return i=t,this.timeline&&this.timeline.revert(t),this.totalTime(-.01,t.suppressEvents),"nested"!==this.data&&Mt(this),i=e,this},e.globalTime=function(t){for(var e=this,r=arguments.length?t:e.rawTime();e;)r=e._start+r/(e._ts||1),e=e._dp;return!this.parent&&this.vars.immediateRender?-1:r},e.repeat=function(t){return arguments.length?(this._repeat=t===1/0?-2:t,Yt(this)):-2===this._repeat?1/0:this._repeat},e.repeatDelay=function(t){if(arguments.length){var e=this._time;return this._rDelay=t,Yt(this),e?this.time(e):this}return this._rDelay},e.yoyo=function(t){return arguments.length?(this._yoyo=t,this):this._yoyo},e.seek=function(t,e){return this.totalTime(Xt(this,t),P(e))},e.restart=function(t,e){return this.play().totalTime(t?-this._delay:0,P(e))},e.play=function(t,e){return null!=t&&this.seek(t,e),this.reversed(!1).paused(!1)},e.reverse=function(t,e){return null!=t&&this.seek(t||this.totalDuration(),e),this.reversed(!0).paused(!1)},e.pause=function(t,e){return null!=t&&this.seek(t,e),this.paused(!0)},e.resume=function(){return this.paused(!1)},e.reversed=function(t){return arguments.length?(!!t!==this.reversed()&&this.timeScale(-this._rts||(t?-1e-8:0)),this):this._rts<0},e.invalidate=function(){return this._initted=this._act=0,this._zTime=-1e-8,this},e.isActive=function(){var t,e=this.parent||this._dp,r=this._start;return!(e&&!(this._ts&&this._initted&&e.isActive()&&(t=e.rawTime(!0))>=r&&t<this.endTime(!0)-1e-8))},e.eventCallback=function(t,e,r){var i=this.vars;return arguments.length>1?(e?(i[t]=e,r&&(i[t+"Params"]=r),"onUpdate"===t&&(this._onUpdate=e)):delete i[t],this):i[t]},e.then=function(t){var e=this;return new Promise((function(r){var i=k(t)?t:mt,n=function(){var t=e.then;e.then=null,k(i)&&(i=i(e))&&(i.then||i===e)&&(e.then=t),r(i),e.then=t};e._initted&&1===e.totalProgress()&&e._ts>=0||!e._tTime&&e._ts<0?n():e._prom=n}))},e.kill=function(){oe(this)},t}();gt(Se.prototype,{_time:0,_start:0,_end:0,_tTime:0,_tDur:0,_dirty:0,_repeat:0,_yoyo:!1,parent:null,_initted:!1,_rDelay:0,_ts:1,_dp:0,ratio:0,_zTime:-1e-8,_prom:0,_ps:!1,_rts:1});var Ee=function(r){function n(e,i){var n;return void 0===e&&(e={}),(n=r.call(this,e)||this).labels={},n.smoothChildTiming=!!e.smoothChildTiming,n.autoRemoveChildren=!!e.autoRemoveChildren,n._sort=P(e.sortChildren),s&&zt(e.parent||s,t(n),i),e.reversed&&n.reverse(),e.paused&&n.paused(!0),e.scrollTrigger&&Ft(t(n),e.scrollTrigger),n}e(n,r);var a=n.prototype;return a.to=function(t,e,r){return Nt(0,arguments,this),this},a.from=function(t,e,r){return Nt(1,arguments,this),this},a.fromTo=function(t,e,r,i){return Nt(2,arguments,this),this},a.set=function(t,e,r){return e.duration=0,e.parent=this,Tt(e).repeatDelay||(e.repeat=0),e.immediateRender=!!e.immediateRender,new qe(t,e,Xt(this,r),1),this},a.call=function(t,e,r){return zt(this,qe.delayedCall(0,t,e),r)},a.staggerTo=function(t,e,r,i,n,s,a){return r.duration=e,r.stagger=r.stagger||i,r.onComplete=s,r.onCompleteParams=a,r.parent=this,new qe(t,r,Xt(this,n)),this},a.staggerFrom=function(t,e,r,i,n,s,a){return r.runBackwards=1,Tt(r).immediateRender=P(r.immediateRender),this.staggerTo(t,e,r,i,n,s,a)},a.staggerFromTo=function(t,e,r,i,n,s,a,o){return i.startAt=r,Tt(i).immediateRender=P(i.immediateRender),this.staggerTo(t,e,i,n,s,a,o)},a.render=function(t,e,r){var n,a,o,u,h,f,l,c,p,_,d,m,g=this._time,v=this._dirty?this.totalDuration():this._tDur,y=this._dur,x=t<=0?0:ft(t),T=this._zTime<0!=t<0&&(this._initted||!y);if(this!==s&&x>v&&t>=0&&(x=v),x!==this._tTime||r||T){if(g!==this._time&&y&&(x+=this._time-g,t+=this._time-g),n=x,p=this._start,f=!(c=this._ts),T&&(y||(g=this._zTime),(t||!e)&&(this._zTime=t)),this._repeat){if(d=this._yoyo,h=y+this._rDelay,this._repeat<-1&&t<0)return this.totalTime(100*h+t,e,r);if(n=ft(x%h),x===v?(u=this._repeat,n=y):((u=~~(x/h))&&u===x/h&&(n=y,u--),n>y&&(n=y)),_=Dt(this._tTime,h),!g&&this._tTime&&_!==u&&(_=u),d&&1&u&&(n=y-n,m=1),u!==_&&!this._lock){var w=d&&1&_,b=w===(d&&1&u);if(u<_&&(w=!w),g=w?0:y,this._lock=1,this.render(g||(m?0:ft(u*h)),e,!y)._lock=0,this._tTime=x,!e&&this.parent&&ae(this,"onRepeat"),this.vars.repeatRefresh&&!m&&(this.invalidate()._lock=1),g&&g!==this._time||f!==!this._ts||this.vars.onRepeat&&!this.parent&&!this._act)return this;if(y=this._dur,v=this._tDur,b&&(this._lock=2,g=w?y:-1e-4,this.render(g,!0),this.vars.repeatRefresh&&!m&&this.invalidate()),this._lock=0,!this._ts&&!f)return this;Me(this,m)}}if(this._hasPause&&!this._forcing&&this._lock<2&&(l=function(t,e,r){var i;if(r>e)for(i=t._first;i&&i._start<=r;){if("isPause"===i.data&&i._start>e)return i;i=i._next}else for(i=t._last;i&&i._start>=r;){if("isPause"===i.data&&i._start<e)return i;i=i._prev}}(this,ft(g),ft(n)))&&(x-=n-(n=l._start)),this._tTime=x,this._time=n,this._act=!c,this._initted||(this._onUpdate=this.vars.onUpdate,this._initted=1,this._zTime=t,g=0),!g&&n&&!e&&(ae(this,"onStart"),this._tTime!==x))return this;if(n>=g&&t>=0)for(a=this._first;a;){if(o=a._next,(a._act||n>=a._start)&&a._ts&&l!==a){if(a.parent!==this)return this.render(t,e,r);if(a.render(a._ts>0?(n-a._start)*a._ts:(a._dirty?a.totalDuration():a._tDur)+(n-a._start)*a._ts,e,r),n!==this._time||!this._ts&&!f){l=0,o&&(x+=this._zTime=-1e-8);break}}a=o}else{r=r||i,a=this._last;for(var M=t<0?t:n;a;){if(o=a._prev,(a._act||M<=a._end)&&a._ts&&l!==a){if(a.parent!==this)return this.render(t,e,r);if(a.render(a._ts>0?(M-a._start)*a._ts:(a._dirty?a.totalDuration():a._tDur)+(M-a._start)*a._ts,e,r),n!==this._time||!this._ts&&!f){l=0,o&&(x+=this._zTime=M?-1e-8:1e-8);break}}a=o}}if(l&&!e&&(this.pause(),l.render(n>=g?0:-1e-8)._zTime=n>=g?1:-1,this._ts))return this._start=p,St(this),this.render(t,e,r);this._onUpdate&&!e&&ae(this,"onUpdate",!0),(x===v&&this._tTime>=this.totalDuration()||!x&&g)&&(p!==this._start&&Math.abs(c)===Math.abs(this._ts)||this._lock||((t||!y)&&(x===v&&this._ts>0||!x&&this._ts<0)&&Mt(this,1),e||t<0&&!g||!x&&!g&&v||(ae(this,x===v&&t>=0?"onComplete":"onReverseComplete",!0),this._prom&&!(x<v&&this.timeScale()>0)&&this._prom())))}return this},a.add=function(t,e){var r=this;if(C(e)||(e=Xt(this,e,t)),!(t instanceof Se)){if(z(t))return t.forEach((function(t){return r.add(t,e)})),this;if(O(t))return this.addLabel(t,e);if(!k(t))return this;t=qe.delayedCall(0,t)}return this!==t?zt(this,t,e):this},a.getChildren=function(t,e,r,i){void 0===t&&(t=!0),void 0===e&&(e=!0),void 0===r&&(r=!0),void 0===i&&(i=-1e8);for(var n=[],s=this._first;s;)s._start>=i&&(s instanceof qe?e&&n.push(s):(r&&n.push(s),t&&n.push.apply(n,s.getChildren(!0,e,r)))),s=s._next;return n},a.getById=function(t){for(var e=this.getChildren(1,1,1),r=e.length;r--;)if(e[r].vars.id===t)return e[r]},a.remove=function(t){return O(t)?this.removeLabel(t):k(t)?this.killTweensOf(t):(bt(this,t),t===this._recent&&(this._recent=this._last),Ot(this))},a.totalTime=function(t,e){return arguments.length?(this._forcing=1,!this._dp&&this._ts&&(this._start=ft(ge.time-(this._ts>0?t/this._ts:(this.totalDuration()-t)/-this._ts))),r.prototype.totalTime.call(this,t,e),this._forcing=0,this):this._tTime},a.addLabel=function(t,e){return this.labels[t]=Xt(this,e),this},a.removeLabel=function(t){return delete this.labels[t],this},a.addPause=function(t,e,r){var i=qe.delayedCall(0,e||Q,r);return i.data="isPause",this._hasPause=1,zt(this,i,Xt(this,t))},a.removePause=function(t){var e=this._first;for(t=Xt(this,t);e;)e._start===t&&"isPause"===e.data&&Mt(e),e=e._next},a.killTweensOf=function(t,e,r){for(var i=this.getTweensOf(t,r),n=i.length;n--;)Re!==i[n]&&i[n].kill(t,e);return this},a.getTweensOf=function(t,e){for(var r,i=[],n=Ht(t),s=this._first,a=C(e);s;)s instanceof qe?ct(s._targets,n)&&(a?(!Re||s._initted&&s._ts)&&s.globalTime(0)<=e&&s.globalTime(s.totalDuration())>e:!e||s.isActive())&&i.push(s):(r=s.getTweensOf(n,e)).length&&i.push.apply(i,r),s=s._next;return i},a.tweenTo=function(t,e){e=e||{};var r,i=this,n=Xt(i,t),s=e,a=s.startAt,o=s.onStart,u=s.onStartParams,h=s.immediateRender,f=qe.to(i,gt({ease:e.ease||"none",lazy:!1,immediateRender:!1,time:n,overwrite:"auto",duration:e.duration||Math.abs((n-(a&&"time"in a?a.time:i._time))/i.timeScale())||1e-8,onStart:function(){if(i.pause(),!r){var t=e.duration||Math.abs((n-(a&&"time"in a?a.time:i._time))/i.timeScale());f._dur!==t&&It(f,t,0,1).render(f._time,!0,!0),r=1}o&&o.apply(f,u||[])}},e));return h?f.render(0):f},a.tweenFromTo=function(t,e,r){return this.tweenTo(e,gt({startAt:{time:Xt(this,t)}},r))},a.recent=function(){return this._recent},a.nextLabel=function(t){return void 0===t&&(t=this._time),se(this,Xt(this,t))},a.previousLabel=function(t){return void 0===t&&(t=this._time),se(this,Xt(this,t),1)},a.currentLabel=function(t){return arguments.length?this.seek(t,!0):this.previousLabel(this._time+1e-8)},a.shiftChildren=function(t,e,r){void 0===r&&(r=0);for(var i,n=this._first,s=this.labels;n;)n._start>=r&&(n._start+=t,n._end+=t),n=n._next;if(e)for(i in s)s[i]>=r&&(s[i]+=t);return Ot(this)},a.invalidate=function(){var t=this._first;for(this._lock=0;t;)t.invalidate(),t=t._next;return r.prototype.invalidate.call(this)},a.clear=function(t){void 0===t&&(t=!0);for(var e,r=this._first;r;)e=r._next,this.remove(r),r=e;return this._dp&&(this._time=this._tTime=this._pTime=0),t&&(this.labels={}),Ot(this)},a.totalDuration=function(t){var e,r,i,n=0,a=this,o=a._last,u=1e8;if(arguments.length)return a.timeScale((a._repeat<0?a.duration():a.totalDuration())/(a.reversed()?-t:t));if(a._dirty){for(i=a.parent;o;)e=o._prev,o._dirty&&o.totalDuration(),(r=o._start)>u&&a._sort&&o._ts&&!a._lock?(a._lock=1,zt(a,o,r-o._delay,1)._lock=0):u=r,r<0&&o._ts&&(n-=r,(!i&&!a._dp||i&&i.smoothChildTiming)&&(a._start+=r/a._ts,a._time-=r,a._tTime-=r),a.shiftChildren(-r,!1,-Infinity),u=0),o._end>n&&o._ts&&(n=o._end),o=e;It(a,a===s&&a._time>n?a._time:n,1,1),a._dirty=0}return a._tDur},n.updateRoot=function(t){if(s._ts&&(_t(s,Pt(t,s)),f=ge.frame),ge.frame>=rt){rt+=g.autoSleep||120;var e=s._first;if((!e||!e._ts)&&g.autoSleep&&ge._listeners.length<2){for(;e&&!e._ts;)e=e._next;e||ge.sleep()}}},n}(Se);gt(Ee.prototype,{_lock:0,_hasPause:0,_forcing:0});var Re,ze,Fe=function(t,e,r,i,n,s,a){var o,u,h,f,l,c,p,_,d=new ir(this._pt,t,e,0,1,$e,null,n),m=0,g=0;for(d.b=r,d.e=i,r+="",(p=~(i+="").indexOf("random("))&&(i=ie(i)),s&&(s(_=[r,i],t,e),r=_[0],i=_[1]),u=r.match(I)||[];o=I.exec(i);)f=o[0],l=i.substring(m,o.index),h?h=(h+1)%5:"rgba("===l.substr(-5)&&(h=1),f!==u[g++]&&(c=parseFloat(u[g-1])||0,d._pt={_next:d._pt,p:l||1===g?l:",",s:c,c:"="===f.charAt(1)?lt(c,f)-c:parseFloat(f)-c,m:h&&h<4?Math.round:0},m=I.lastIndex);return d.c=m<i.length?i.substring(m,i.length):"",d.fp=a,(Y.test(i)||p)&&(d.e=0),this._pt=d,d},Be=function(t,e,r,i,n,s,a,o,u,h){k(i)&&(i=i(n||0,t,s));var f,l=t[e],c="get"!==r?r:k(l)?u?t[e.indexOf("set")||!k(t["get"+e.substr(3)])?e:"get"+e.substr(3)](u):t[e]():l,p=k(l)?u?Ge:je:Ve;if(O(i)&&(~i.indexOf("random(")&&(i=ie(i)),"="===i.charAt(1)&&((f=lt(c,i)+(jt(c)||0))||0===f)&&(i=f)),!h||c!==i||ze)return isNaN(c*i)||""===i?(!l&&!(e in t)&&j(e,i),Fe.call(this,t,e,c,i,p,o||g.stringFilter,u)):(f=new ir(this._pt,t,e,+c||0,i-(c||0),"boolean"==typeof l?Ze:He,0,p),u&&(f.fp=u),a&&f.modifier(a,this,t),this._pt=f)},Le=function(t,e,r,i,n,s){var a,o,u,h;if(tt[t]&&!1!==(a=new tt[t]).init(n,a.rawVars?e[t]:function(t,e,r,i,n){if(k(t)&&(t=Ue(t,n,e,r,i)),!D(t)||t.style&&t.nodeType||z(t)||R(t))return O(t)?Ue(t,n,e,r,i):t;var s,a={};for(s in t)a[s]=Ue(t[s],n,e,r,i);return a}(e[t],i,n,s,r),r,i,s)&&(r._pt=o=new ir(r._pt,n,t,0,1,a.render,a,0,a.priority),r!==l))for(u=r._ptLookup[r._targets.indexOf(n)],h=a._props.length;h--;)u[a._props[h]]=o;return a},Ie=function t(e,n){var a,o,u,h,f,l,c,p,_,d,m,g,y,x=e.vars,T=x.ease,w=x.startAt,b=x.immediateRender,M=x.lazy,O=x.onUpdate,k=x.onUpdateParams,C=x.callbackScope,A=x.runBackwards,D=x.yoyoEase,S=x.keyframes,E=x.autoRevert,R=e._dur,z=e._startAt,F=e._targets,B=e.parent,L=B&&"nested"===B.data?B.parent._targets:F,I="auto"===e._overwrite&&!r,Y=e.timeline;if(Y&&(!S||!T)&&(T="none"),e._ease=Oe(T,v.ease),e._yEase=D?be(Oe(!0===D?T:D,v.ease)):0,D&&e._yoyo&&!e._repeat&&(D=e._yEase,e._yEase=e._ease,e._ease=D),e._from=!Y&&!!x.runBackwards,!Y||S&&!x.stagger){if(g=(p=F[0]?at(F[0]).harness:0)&&x[p.prop],a=xt(x,$),z&&(z.revert(A&&R?Z:H),z._lazy=0),w){if(Mt(e._startAt=qe.set(F,gt({data:"isStart",overwrite:!1,parent:B,immediateRender:!0,lazy:P(M),startAt:null,delay:0,onUpdate:O,onUpdateParams:k,callbackScope:C,stagger:0},w))),n<0&&(i||!b&&!E)&&e._startAt.revert(Z),b&&R&&n<=0)return void(n&&(e._zTime=n))}else if(A&&R&&!z)if(n&&(b=!1),u=gt({overwrite:!1,data:"isFromStart",lazy:b&&P(M),immediateRender:b,stagger:0,parent:B},a),g&&(u[p.prop]=g),Mt(e._startAt=qe.set(F,u)),n<0&&(i?e._startAt.revert(Z):e._startAt.render(-1,!0)),e._zTime=n,b){if(!n)return}else t(e._startAt,1e-8);for(e._pt=e._ptCache=0,M=R&&P(M)||M&&!R,o=0;o<F.length;o++){if(c=(f=F[o])._gsap||st(F)[o]._gsap,e._ptLookup[o]=d={},K[c.id]&&J.length&&pt(),m=L===F?o:L.indexOf(f),p&&!1!==(_=new p).init(f,g||a,e,m,L)&&(e._pt=h=new ir(e._pt,f,_.name,0,1,_.render,_,0,_.priority),_._props.forEach((function(t){d[t]=h})),_.priority&&(l=1)),!p||g)for(u in a)tt[u]&&(_=Le(u,a,e,m,f,L))?_.priority&&(l=1):d[u]=h=Be.call(e,f,u,"get",a[u],m,L,0,x.stringFilter);e._op&&e._op[o]&&e.kill(f,e._op[o]),I&&e._pt&&(Re=e,s.killTweensOf(f,d,e.globalTime(n)),y=!e.parent,Re=0),e._pt&&M&&(K[c.id]=1)}l&&rr(e),e._onInit&&e._onInit(e)}e._onUpdate=O,e._initted=(!e._op||e._pt)&&!y,S&&n<=0&&Y.render(1e8,!0,!0)},Ye=function(t,e,r,i){var n,s,a=e.ease||i||"power1.inOut";if(z(e))s=r[t]||(r[t]=[]),e.forEach((function(t,r){return s.push({t:r/(e.length-1)*100,v:t,e:a})}));else for(n in e)s=r[n]||(r[n]=[]),"ease"===n||s.push({t:parseFloat(t),v:e[n],e:a})},Ue=function(t,e,r,i,n){return k(t)?t.call(e,r,i,n):O(t)&&~t.indexOf("random(")?ie(t):t},Xe=nt+"repeat,repeatDelay,yoyo,repeatRefresh,yoyoEase,autoRevert",Ne={};ut(Xe+",id,stagger,delay,duration,paused,scrollTrigger",(function(t){return Ne[t]=1}));var qe=function(n){function a(e,i,a,o){var u;"number"==typeof i&&(a.duration=i,i=a,a=null);var h,f,l,c,p,_,d,m,v=(u=n.call(this,o?i:Tt(i))||this).vars,y=v.duration,x=v.delay,T=v.immediateRender,w=v.stagger,b=v.overwrite,M=v.keyframes,O=v.defaults,k=v.scrollTrigger,A=v.yoyoEase,S=i.parent||s,F=(z(e)||R(e)?C(e[0]):"length"in i)?[e]:Ht(e);if(u._targets=F.length?st(F):G("GSAP target "+e+" not found. https://greensock.com",!g.nullTargetWarn)||[],u._ptLookup=[],u._overwrite=b,M||w||E(y)||E(x)){if(i=u.vars,(h=u.timeline=new Ee({data:"nested",defaults:O||{}})).kill(),h.parent=h._dp=t(u),h._start=0,w||E(y)||E(x)){if(c=F.length,d=w&&Jt(w),D(w))for(p in w)~Xe.indexOf(p)&&(m||(m={}),m[p]=w[p]);for(f=0;f<c;f++)(l=xt(i,Ne)).stagger=0,A&&(l.yoyoEase=A),m&&vt(l,m),_=F[f],l.duration=+Ue(y,t(u),f,_,F),l.delay=(+Ue(x,t(u),f,_,F)||0)-u._delay,!w&&1===c&&l.delay&&(u._delay=x=l.delay,u._start+=x,l.delay=0),h.to(_,l,d?d(f,_,F):0),h._ease=ye.none;h.duration()?y=x=0:u.timeline=0}else if(M){Tt(gt(h.vars.defaults,{ease:"none"})),h._ease=Oe(M.ease||i.ease||"none");var B,L,I,Y=0;if(z(M))M.forEach((function(t){return h.to(F,t,">")})),h.duration();else{for(p in l={},M)"ease"===p||"easeEach"===p||Ye(p,M[p],l,M.easeEach);for(p in l)for(B=l[p].sort((function(t,e){return t.t-e.t})),Y=0,f=0;f<B.length;f++)(I={ease:(L=B[f]).e,duration:(L.t-(f?B[f-1].t:0))/100*y})[p]=L.v,h.to(F,I,Y),Y+=I.duration;h.duration()<y&&h.to({},{duration:y-h.duration()})}}y||u.duration(y=h.duration())}else u.timeline=0;return!0!==b||r||(Re=t(u),s.killTweensOf(F),Re=0),zt(S,t(u),a),i.reversed&&u.reverse(),i.paused&&u.paused(!0),(T||!y&&!M&&u._start===ft(S._time)&&P(T)&&function t(e){return!e||e._ts&&t(e.parent)}(t(u))&&"nested"!==S.data)&&(u._tTime=-1e-8,u.render(Math.max(0,-x))),k&&Ft(t(u),k),u}e(a,n);var o=a.prototype;return o.render=function(t,e,r){var n,s,a,o,u,h,f,l,c,p=this._time,_=this._tDur,d=this._dur,m=t<0,g=t>_-1e-8&&!m?_:t<1e-8?0:t;if(d){if(g!==this._tTime||!t||r||!this._initted&&this._tTime||this._startAt&&this._zTime<0!==m){if(n=g,l=this.timeline,this._repeat){if(o=d+this._rDelay,this._repeat<-1&&m)return this.totalTime(100*o+t,e,r);if(n=ft(g%o),g===_?(a=this._repeat,n=d):((a=~~(g/o))&&a===g/o&&(n=d,a--),n>d&&(n=d)),(h=this._yoyo&&1&a)&&(c=this._yEase,n=d-n),u=Dt(this._tTime,o),n===p&&!r&&this._initted)return this._tTime=g,this;a!==u&&(l&&this._yEase&&Me(l,h),!this.vars.repeatRefresh||h||this._lock||(this._lock=r=1,this.render(ft(o*a),!0).invalidate()._lock=0))}if(!this._initted){if(Bt(this,m?t:n,r,e))return this._tTime=0,this;if(p!==this._time)return this;if(d!==this._dur)return this.render(t,e,r)}if(this._tTime=g,this._time=n,!this._act&&this._ts&&(this._act=1,this._lazy=0),this.ratio=f=(c||this._ease)(n/d),this._from&&(this.ratio=f=1-f),n&&!p&&!e&&(ae(this,"onStart"),this._tTime!==g))return this;for(s=this._pt;s;)s.r(f,s.d),s=s._next;l&&l.render(t<0?t:!n&&h?-1e-8:l._dur*l._ease(n/this._dur),e,r)||this._startAt&&(this._zTime=t),this._onUpdate&&!e&&(m&&Ct(this,t,0,r),ae(this,"onUpdate")),this._repeat&&a!==u&&this.vars.onRepeat&&!e&&this.parent&&ae(this,"onRepeat"),g!==this._tDur&&g||this._tTime!==g||(m&&!this._onUpdate&&Ct(this,t,0,!0),(t||!d)&&(g===this._tDur&&this._ts>0||!g&&this._ts<0)&&Mt(this,1),e||m&&!p||!g&&!p||(ae(this,g===_?"onComplete":"onReverseComplete",!0),this._prom&&!(g<_&&this.timeScale()>0)&&this._prom()))}}else!function(t,e,r,n){var s,a,o,u=t.ratio,h=e<0||!e&&(!t._start&&function t(e){var r=e.parent;return r&&r._ts&&r._initted&&!r._lock&&(r.rawTime()<0||t(r))}(t)&&(t._initted||!Lt(t))||(t._ts<0||t._dp._ts<0)&&!Lt(t))?0:1,f=t._rDelay,l=0;if(f&&t._repeat&&(l=Vt(0,t._tDur,e),a=Dt(l,f),t._yoyo&&1&a&&(h=1-h),a!==Dt(t._tTime,f)&&(u=1-h,t.vars.repeatRefresh&&t._initted&&t.invalidate())),h!==u||i||n||1e-8===t._zTime||!e&&t._zTime){if(!t._initted&&Bt(t,e,n,r))return;for(o=t._zTime,t._zTime=e||(r?1e-8:0),r||(r=e&&!o),t.ratio=h,t._from&&(h=1-h),t._time=0,t._tTime=l,s=t._pt;s;)s.r(h,s.d),s=s._next;e<0&&Ct(t,e,0,!0),t._onUpdate&&!r&&ae(t,"onUpdate"),l&&t._repeat&&!r&&t.parent&&ae(t,"onRepeat"),(e>=t._tDur||e<0)&&t.ratio===h&&(h&&Mt(t,1),r||(ae(t,h?"onComplete":"onReverseComplete",!0),t._prom&&t._prom()))}else t._zTime||(t._zTime=e)}(this,t,e,r);return this},o.targets=function(){return this._targets},o.invalidate=function(){return this._pt=this._op=this._startAt=this._onUpdate=this._lazy=this.ratio=0,this._ptLookup=[],this.timeline&&this.timeline.invalidate(),n.prototype.invalidate.call(this)},o.resetTo=function(t,e,r,i){c||ge.wake(),this._ts||this.play();var n=Math.min(this._dur,(this._dp._time-this._start)*this._ts);return this._initted||Ie(this,n),function(t,e,r,i,n,s,a){var o,u,h,f,l=(t._pt&&t._ptCache||(t._ptCache={}))[e];if(!l)for(l=t._ptCache[e]=[],h=t._ptLookup,f=t._targets.length;f--;){if((o=h[f][e])&&o.d&&o.d._pt)for(o=o.d._pt;o&&o.p!==e&&o.fp!==e;)o=o._next;if(!o)return ze=1,t.vars[e]="+=0",Ie(t,a),ze=0,1;l.push(o)}for(f=l.length;f--;)(o=(u=l[f])._pt||u).s=!i&&0!==i||n?o.s+(i||0)+s*o.c:i,o.c=r-o.s,u.e&&(u.e=ht(r)+jt(u.e)),u.b&&(u.b=o.s+jt(u.b))}(this,t,e,r,i,this._ease(n/this._dur),n)?this.resetTo(t,e,r,i):(Et(this,0),this.parent||wt(this._dp,this,"_first","_last",this._dp._sort?"_start":0),this.render(0))},o.kill=function(t,e){if(void 0===e&&(e="all"),!(t||e&&"all"!==e))return this._lazy=this._pt=0,this.parent?oe(this):this;if(this.timeline){var r=this.timeline.totalDuration();return this.timeline.killTweensOf(t,e,Re&&!0!==Re.vars.overwrite)._first||oe(this),this.parent&&r!==this.timeline.totalDuration()&&It(this,this._dur*this.timeline._tDur/r,0,1),this}var i,n,s,a,o,u,h,f=this._targets,l=t?Ht(t):f,c=this._ptLookup,p=this._pt;if((!e||"all"===e)&&function(t,e){for(var r=t.length,i=r===e.length;i&&r--&&t[r]===e[r];);return r<0}(f,l))return"all"===e&&(this._pt=0),oe(this);for(i=this._op=this._op||[],"all"!==e&&(O(e)&&(o={},ut(e,(function(t){return o[t]=1})),e=o),e=function(t,e){var r,i,n,s,a=t[0]?at(t[0]).harness:0,o=a&&a.aliases;if(!o)return e;for(i in r=vt({},e),o)if(i in r)for(n=(s=o[i].split(",")).length;n--;)r[s[n]]=r[i];return r}(f,e)),h=f.length;h--;)if(~l.indexOf(f[h]))for(o in n=c[h],"all"===e?(i[h]=e,a=n,s={}):(s=i[h]=i[h]||{},a=e),a)(u=n&&n[o])&&("kill"in u.d&&!0!==u.d.kill(o)||bt(this,u,"_pt"),delete n[o]),"all"!==s&&(s[o]=1);return this._initted&&!this._pt&&p&&oe(this),this},a.to=function(t,e){return new a(t,e,arguments[2])},a.from=function(t,e){return Nt(1,arguments)},a.delayedCall=function(t,e,r,i){return new a(e,0,{immediateRender:!1,lazy:!1,overwrite:!1,delay:t,onComplete:e,onReverseComplete:e,onCompleteParams:r,onReverseCompleteParams:r,callbackScope:i})},a.fromTo=function(t,e,r){return Nt(2,arguments)},a.set=function(t,e){return e.duration=0,e.repeatDelay||(e.repeat=0),new a(t,e)},a.killTweensOf=function(t,e,r){return s.killTweensOf(t,e,r)},a}(Se);gt(qe.prototype,{_targets:[],_lazy:0,_startAt:0,_op:0,_onInit:0}),ut("staggerTo,staggerFrom,staggerFromTo",(function(t){qe[t]=function(){var e=new Ee,r=Gt.call(arguments,0);return r.splice("staggerFromTo"===t?5:4,0,0),e[t].apply(e,r)}}));var Ve=function(t,e,r){return t[e]=r},je=function(t,e,r){return t[e](r)},Ge=function(t,e,r,i){return t[e](i.fp,r)},We=function(t,e,r){return t.setAttribute(e,r)},Qe=function(t,e){return k(t[e])?je:A(t[e])&&t.setAttribute?We:Ve},He=function(t,e){return e.set(e.t,e.p,Math.round(1e6*(e.s+e.c*t))/1e6,e)},Ze=function(t,e){return e.set(e.t,e.p,!!(e.s+e.c*t),e)},$e=function(t,e){var r=e._pt,i="";if(!t&&e.b)i=e.b;else if(1===t&&e.e)i=e.e;else{for(;r;)i=r.p+(r.m?r.m(r.s+r.c*t):Math.round(1e4*(r.s+r.c*t))/1e4)+i,r=r._next;i+=e.c}e.set(e.t,e.p,i,e)},Je=function(t,e){for(var r=e._pt;r;)r.r(t,r.d),r=r._next},Ke=function(t,e,r,i){for(var n,s=this._pt;s;)n=s._next,s.p===i&&s.modifier(t,e,r),s=n},tr=function(t){for(var e,r,i=this._pt;i;)r=i._next,i.p===t&&!i.op||i.op===t?bt(this,i,"_pt"):i.dep||(e=1),i=r;return!e},er=function(t,e,r,i){i.mSet(t,e,i.m.call(i.tween,r,i.mt),i)},rr=function(t){for(var e,r,i,n,s=t._pt;s;){for(e=s._next,r=i;r&&r.pr>s.pr;)r=r._next;(s._prev=r?r._prev:n)?s._prev._next=s:i=s,(s._next=r)?r._prev=s:n=s,s=e}t._pt=i},ir=function(){function t(t,e,r,i,n,s,a,o,u){this.t=e,this.s=i,this.c=n,this.p=r,this.r=s||He,this.d=a||this,this.set=o||Ve,this.pr=u||0,this._next=t,t&&(t._prev=this)}return t.prototype.modifier=function(t,e,r){this.mSet=this.mSet||this.set,this.set=er,this.m=t,this.mt=r,this.tween=e},t}();ut(nt+"parent,duration,ease,delay,overwrite,runBackwards,startAt,yoyo,immediateRender,repeat,repeatDelay,data,paused,reversed,lazy,callbackScope,stringFilter,id,yoyoEase,stagger,inherit,repeatRefresh,keyframes,autoRevert,scrollTrigger",(function(t){return $[t]=1})),N.TweenMax=N.TweenLite=qe,N.TimelineLite=N.TimelineMax=Ee,s=new Ee({sortChildren:!1,defaults:v,autoRemoveChildren:!0,id:"root",smoothChildTiming:!0}),g.stringFilter=me;var nr=[],sr={},ar=[],or=0,ur=function(t){return(sr[t]||ar).map((function(t){return t()}))},hr=function(){var t=Date.now(),e=[];t-or>2&&(ur("matchMediaInit"),nr.forEach((function(t){var r,i,n,s,o=t.queries,u=t.conditions;for(i in o)(r=a.matchMedia(o[i]).matches)&&(n=1),r!==u[i]&&(u[i]=r,s=1);s&&(t.revert(),n&&e.push(t))})),ur("matchMediaRevert"),e.forEach((function(t){return t.onMatch(t)})),or=t,ur("matchMedia"))},fr=function(){function t(t,e){this.selector=e&&Zt(e),this.data=[],this._r=[],this.isReverted=!1,t&&this.add(t)}var e=t.prototype;return e.add=function(t,e,r){k(t)&&(r=e,e=t,t=k);var i=this,s=function(){var t,s=n,a=i.selector;return s&&s.data.push(i),r&&(i.selector=Zt(r)),n=i,t=e.apply(i,arguments),k(t)&&i._r.push(t),n=s,i.selector=a,i.isReverted=!1,t};return i.last=s,t===k?s(i):t?i[t]=s:s},e.ignore=function(t){var e=n;n=null,t(this),n=e},e.getTweens=function(){var e=[];return this.data.forEach((function(r){return r instanceof t?e.push.apply(e,r.getTweens()):r instanceof qe&&r._targets[0]!==r.vars.onComplete&&e.push(r)})),e},e.clear=function(){this._r.length=this.data.length=0},e.kill=function(t,e){var r=this;if(t?(this.getTweens().map((function(t){return{g:t.globalTime(0),t:t}})).sort((function(t,e){return e.g-t.g||-1})).forEach((function(e){return e.t.revert(t)})),this.data.forEach((function(e){return!(e instanceof Se)&&e.revert&&e.revert(t)})),this._r.forEach((function(e){return e(t,r)})),this.isReverted=!0):this.data.forEach((function(t){return t.kill&&t.kill()})),this.clear(),e){var i=nr.indexOf(this);~i&&nr.splice(i,1)}},e.revert=function(t){this.kill(t||{})},t}(),lr=function(){function t(t){this.contexts=[],this.scope=t}var e=t.prototype;return e.add=function(t,e,r){D(t)||(t={matches:t});var i,n,s,o=new fr(0,r||this.scope),u=o.conditions={};for(n in this.contexts.push(o),e=o.add("onMatch",e),o.queries=t,t)"all"===n?s=1:(i=a.matchMedia(t[n]))&&(nr.indexOf(o)<0&&nr.push(o),(u[n]=i.matches)&&(s=1),i.addListener?i.addListener(hr):i.addEventListener("change",hr));return s&&e(o),this},e.revert=function(t){this.kill(t||{})},e.kill=function(t){this.contexts.forEach((function(e){return e.kill(t,!0)}))},t}(),cr={registerPlugin:function(){for(var t=arguments.length,e=new Array(t),r=0;r<t;r++)e[r]=arguments[r];e.forEach((function(t){return ue(t)}))},timeline:function(t){return new Ee(t)},getTweensOf:function(t,e){return s.getTweensOf(t,e)},getProperty:function(t,e,r,i){O(t)&&(t=Ht(t)[0]);var n=at(t||{}).get,s=r?mt:dt;return"native"===r&&(r=""),t?e?s((tt[e]&&tt[e].get||n)(t,e,r,i)):function(e,r,i){return s((tt[e]&&tt[e].get||n)(t,e,r,i))}:t},quickSetter:function(t,e,r){if((t=Ht(t)).length>1){var i=t.map((function(t){return dr.quickSetter(t,e,r)})),n=i.length;return function(t){for(var e=n;e--;)i[e](t)}}t=t[0]||{};var s=tt[e],a=at(t),o=a.harness&&(a.harness.aliases||{})[e]||e,u=s?function(e){var i=new s;l._pt=0,i.init(t,r?e+r:e,l,0,[t]),i.render(1,i),l._pt&&Je(1,l)}:a.set(t,o);return s?u:function(e){return u(t,o,r?e+r:e,a,1)}},quickTo:function(t,e,r){var i,n=dr.to(t,vt(((i={})[e]="+=0.1",i.paused=!0,i),r||{})),s=function(t,r,i){return n.resetTo(e,t,r,i)};return s.tween=n,s},isTweening:function(t){return s.getTweensOf(t,!0).length>0},defaults:function(t){return t&&t.ease&&(t.ease=Oe(t.ease,v.ease)),yt(v,t||{})},config:function(t){return yt(g,t||{})},registerEffect:function(t){var e=t.name,r=t.effect,i=t.plugins,n=t.defaults,s=t.extendTimeline;(i||"").split(",").forEach((function(t){return t&&!tt[t]&&!N[t]&&G(e+" effect requires "+t+" plugin.")})),et[e]=function(t,e,i){return r(Ht(t),gt(e||{},n),i)},s&&(Ee.prototype[e]=function(t,r,i){return this.add(et[e](t,D(r)?r:(i=r)&&{},this),i)})},registerEase:function(t,e){ye[t]=Oe(e)},parseEase:function(t,e){return arguments.length?Oe(t,e):ye},getById:function(t){return s.getById(t)},exportRoot:function(t,e){void 0===t&&(t={});var r,i,n=new Ee(t);for(n.smoothChildTiming=P(t.smoothChildTiming),s.remove(n),n._dp=0,n._time=n._tTime=s._time,r=s._first;r;)i=r._next,!e&&!r._dur&&r instanceof qe&&r.vars.onComplete===r._targets[0]||zt(n,r,r._start-r._delay),r=i;return zt(s,n,0),n},context:function(t,e){return t?new fr(t,e):n},matchMedia:function(t){return new lr(t)},matchMediaRefresh:function(){return nr.forEach((function(t){var e,r,i=t.conditions;for(r in i)i[r]&&(i[r]=!1,e=1);e&&t.revert()}))||hr()},addEventListener:function(t,e){var r=sr[t]||(sr[t]=[]);~r.indexOf(e)||r.push(e)},removeEventListener:function(t,e){var r=sr[t],i=r&&r.indexOf(e);i>=0&&r.splice(i,1)},utils:{wrap:function t(e,r,i){var n=r-e;return z(e)?re(e,t(0,e.length),r):qt(i,(function(t){return(n+(t-e)%n)%n+e}))},wrapYoyo:function t(e,r,i){var n=r-e,s=2*n;return z(e)?re(e,t(0,e.length-1),r):qt(i,(function(t){return e+((t=(s+(t-e)%s)%s||0)>n?s-t:t)}))},distribute:Jt,random:ee,snap:te,normalize:function(t,e,r){return ne(t,e,0,1,r)},getUnit:jt,clamp:function(t,e,r){return qt(r,(function(r){return Vt(t,e,r)}))},splitColor:le,toArray:Ht,selector:Zt,mapRange:ne,pipe:function(){for(var t=arguments.length,e=new Array(t),r=0;r<t;r++)e[r]=arguments[r];return function(t){return e.reduce((function(t,e){return e(t)}),t)}},unitize:function(t,e){return function(r){return t(parseFloat(r))+(e||jt(r))}},interpolate:function t(e,r,i,n){var s=isNaN(e+r)?0:function(t){return(1-t)*e+t*r};if(!s){var a,o,u,h,f,l=O(e),c={};if(!0===i&&(n=1)&&(i=null),l)e={p:e},r={p:r};else if(z(e)&&!z(r)){for(u=[],h=e.length,f=h-2,o=1;o<h;o++)u.push(t(e[o-1],e[o]));h--,s=function(t){t*=h;var e=Math.min(f,~~t);return u[e](t-e)},i=r}else n||(e=vt(z(e)?[]:{},e));if(!u){for(a in r)Be.call(c,e,a,"get",r[a]);s=function(t){return Je(t,c)||(l?e.p:e)}}}return qt(i,s)},shuffle:$t},install:V,effects:et,ticker:ge,updateRoot:Ee.updateRoot,plugins:tt,globalTimeline:s,core:{PropTween:ir,globals:W,Tween:qe,Timeline:Ee,Animation:Se,getCache:at,_removeLinkedListItem:bt,reverting:function(){return i},context:function(t){return t&&n&&(n.data.push(t),t._ctx=n),n},suppressOverwrites:function(t){return r=t}}};ut("to,from,fromTo,delayedCall,set,killTweensOf",(function(t){return cr[t]=qe[t]})),ge.add(Ee.updateRoot),l=cr.to({},{duration:0});var pr=function(t,e){for(var r=t._pt;r&&r.p!==e&&r.op!==e&&r.fp!==e;)r=r._next;return r},_r=function(t,e){return{name:t,rawVars:1,init:function(t,r,i){i._onInit=function(t){var i,n;if(O(r)&&(i={},ut(r,(function(t){return i[t]=1})),r=i),e){for(n in i={},r)i[n]=e(r[n]);r=i}!function(t,e){var r,i,n,s=t._targets;for(r in e)for(i=s.length;i--;)(n=t._ptLookup[i][r])&&(n=n.d)&&(n._pt&&(n=pr(n,r)),n&&n.modifier&&n.modifier(e[r],t,s[i],r))}(t,r)}}}},dr=cr.registerPlugin({name:"attr",init:function(t,e,r,i,n){var s,a,o;for(s in this.tween=r,e)o=t.getAttribute(s)||"",(a=this.add(t,"setAttribute",(o||0)+"",e[s],i,n,0,0,s)).op=s,a.b=o,this._props.push(s)},render:function(t,e){for(var r=e._pt;r;)i?r.set(r.t,r.p,r.b,r):r.r(t,r.d),r=r._next}},{name:"endArray",init:function(t,e){for(var r=e.length;r--;)this.add(t,r,t[r]||0,e[r],0,0,0,0,0,1)}},_r("roundProps",Kt),_r("modifiers"),_r("snap",te))||cr;qe.version=Ee.version=dr.version="3.11.0",h=1,S()&&ve(),ye.Power0,ye.Power1,ye.Power2,ye.Power3,ye.Power4,ye.Linear,ye.Quad,ye.Cubic,ye.Quart,ye.Quint,ye.Strong,ye.Elastic,ye.Back,ye.SteppedEase,ye.Bounce,ye.Sine,ye.Expo,ye.Circ;
-/*!
- * CSSPlugin 3.11.0
- * https://greensock.com
- *
- * Copyright 2008-2022, GreenSock. All rights reserved.
- * Subject to the terms at https://greensock.com/standard-license or for
- * Club GreenSock members, the agreement issued with that membership.
- * @author: Jack Doyle, jack@greensock.com
-*/
-var mr,gr,vr,yr,xr,Tr,wr,br,Mr={},Or=180/Math.PI,kr=Math.PI/180,Cr=Math.atan2,Ar=/([A-Z])/g,Dr=/(left|right|width|margin|padding|x)/i,Pr=/[\s,\(]\S/,Sr={autoAlpha:"opacity,visibility",scale:"scaleX,scaleY",alpha:"opacity"},Er=function(t,e){return e.set(e.t,e.p,Math.round(1e4*(e.s+e.c*t))/1e4+e.u,e)},Rr=function(t,e){return e.set(e.t,e.p,1===t?e.e:Math.round(1e4*(e.s+e.c*t))/1e4+e.u,e)},zr=function(t,e){return e.set(e.t,e.p,t?Math.round(1e4*(e.s+e.c*t))/1e4+e.u:e.b,e)},Fr=function(t,e){var r=e.s+e.c*t;e.set(e.t,e.p,~~(r+(r<0?-.5:.5))+e.u,e)},Br=function(t,e){return e.set(e.t,e.p,t?e.e:e.b,e)},Lr=function(t,e){return e.set(e.t,e.p,1!==t?e.b:e.e,e)},Ir=function(t,e,r){return t.style[e]=r},Yr=function(t,e,r){return t.style.setProperty(e,r)},Ur=function(t,e,r){return t._gsap[e]=r},Xr=function(t,e,r){return t._gsap.scaleX=t._gsap.scaleY=r},Nr=function(t,e,r,i,n){var s=t._gsap;s.scaleX=s.scaleY=r,s.renderTransform(n,s)},qr=function(t,e,r,i,n){var s=t._gsap;s[e]=r,s.renderTransform(n,s)},Vr="transform",jr=Vr+"Origin",Gr=function(t){var e=this,r=this.target,i=r.style;if(t in Mr){if(this.tfm=this.tfm||{},"transform"!==t&&(~(t=Sr[t]||t).indexOf(",")?t.split(",").forEach((function(t){return e.tfm[t]=fi(r,t)})):this.tfm[t]=r._gsap.x?r._gsap[t]:fi(r,t)),r._gsap.svg&&(this.svg=r.getAttribute(t)||""),this.props.indexOf(Vr)>=0)return;t=Vr}i&&this.props.push(t,i[t])},Wr=function(t){t.translate&&(t.removeProperty("translate"),t.removeProperty("scale"),t.removeProperty("rotate"))},Qr=function(){var t,e,r=this.props,i=this.target,n=i.style,s=i._gsap;for(t=0;t<r.length;t+=2)r[t+1]?n[r[t]]=r[t+1]:n.removeProperty(r[t].replace(Ar,"-$1").toLowerCase());if(this.tfm){for(e in s.svg&&i.setAttribute("transform",this.svg||""),this.tfm)s[e]=this.tfm[e];!(t=wr())||t.isStart||n[Vr]||(Wr(n),s.uncache=1)}},Hr=function(t,e){var r={target:t,props:[],revert:Qr,save:Gr};return e&&e.split(",").forEach((function(t){return r.save(t)})),r},Zr=function(t,e){var r=gr.createElementNS?gr.createElementNS((e||"http://www.w3.org/1999/xhtml").replace(/^https/,"http"),t):gr.createElement(t);return r.style?r:gr.createElement(t)},$r=function t(e,r,i){var n=getComputedStyle(e);return n[r]||n.getPropertyValue(r.replace(Ar,"-$1").toLowerCase())||n.getPropertyValue(r)||!i&&t(e,Kr(r)||r,1)||""},Jr="O,Moz,ms,Ms,Webkit".split(","),Kr=function(t,e,r){var i=(e||xr).style,n=5;if(t in i&&!r)return t;for(t=t.charAt(0).toUpperCase()+t.substr(1);n--&&!(Jr[n]+t in i););return n<0?null:(3===n?"ms":n>=0?Jr[n]:"")+t},ti=function(){"undefined"!=typeof window&&window.document&&(mr=window,gr=mr.document,vr=gr.documentElement,xr=Zr("div")||{style:{}},Zr("div"),Vr=Kr(Vr),jr=Vr+"Origin",xr.style.cssText="border-width:0;line-height:0;position:absolute;padding:0",br=!!Kr("perspective"),wr=dr.core.reverting,yr=1)},ei=function t(e){var r,i=Zr("svg",this.ownerSVGElement&&this.ownerSVGElement.getAttribute("xmlns")||"http://www.w3.org/2000/svg"),n=this.parentNode,s=this.nextSibling,a=this.style.cssText;if(vr.appendChild(i),i.appendChild(this),this.style.display="block",e)try{r=this.getBBox(),this._gsapBBox=this.getBBox,this.getBBox=t}catch(t){}else this._gsapBBox&&(r=this._gsapBBox());return n&&(s?n.insertBefore(this,s):n.appendChild(this)),vr.removeChild(i),this.style.cssText=a,r},ri=function(t,e){for(var r=e.length;r--;)if(t.hasAttribute(e[r]))return t.getAttribute(e[r])},ii=function(t){var e;try{e=t.getBBox()}catch(r){e=ei.call(t,!0)}return e&&(e.width||e.height)||t.getBBox===ei||(e=ei.call(t,!0)),!e||e.width||e.x||e.y?e:{x:+ri(t,["x","cx","x1"])||0,y:+ri(t,["y","cy","y1"])||0,width:0,height:0}},ni=function(t){return!(!t.getCTM||t.parentNode&&!t.ownerSVGElement||!ii(t))},si=function(t,e){if(e){var r=t.style;e in Mr&&e!==jr&&(e=Vr),r.removeProperty?("ms"!==e.substr(0,2)&&"webkit"!==e.substr(0,6)||(e="-"+e),r.removeProperty(e.replace(Ar,"-$1").toLowerCase())):r.removeAttribute(e)}},ai=function(t,e,r,i,n,s){var a=new ir(t._pt,e,r,0,1,s?Lr:Br);return t._pt=a,a.b=i,a.e=n,t._props.push(r),a},oi={deg:1,rad:1,turn:1},ui={grid:1,flex:1},hi=function t(e,r,i,n){var s,a,o,u,h=parseFloat(i)||0,f=(i+"").trim().substr((h+"").length)||"px",l=xr.style,c=Dr.test(r),p="svg"===e.tagName.toLowerCase(),_=(p?"client":"offset")+(c?"Width":"Height"),d="px"===n,m="%"===n;return n===f||!h||oi[n]||oi[f]?h:("px"!==f&&!d&&(h=t(e,r,i,"px")),u=e.getCTM&&ni(e),!m&&"%"!==f||!Mr[r]&&!~r.indexOf("adius")?(l[c?"width":"height"]=100+(d?f:n),a=~r.indexOf("adius")||"em"===n&&e.appendChild&&!p?e:e.parentNode,u&&(a=(e.ownerSVGElement||{}).parentNode),a&&a!==gr&&a.appendChild||(a=gr.body),(o=a._gsap)&&m&&o.width&&c&&o.time===ge.time&&!o.uncache?ht(h/o.width*100):((m||"%"===f)&&!ui[$r(a,"display")]&&(l.position=$r(e,"position")),a===e&&(l.position="static"),a.appendChild(xr),s=xr[_],a.removeChild(xr),l.position="absolute",c&&m&&((o=at(a)).time=ge.time,o.width=a[_]),ht(d?s*h/100:s&&h?100/s*h:0))):(s=u?e.getBBox()[c?"width":"height"]:e[_],ht(m?h/s*100:h/100*s)))},fi=function(t,e,r,i){var n;return yr||ti(),e in Sr&&"transform"!==e&&~(e=Sr[e]).indexOf(",")&&(e=e.split(",")[0]),Mr[e]&&"transform"!==e?(n=Ti(t,i),n="transformOrigin"!==e?n[e]:n.svg?n.origin:wi($r(t,jr))+" "+n.zOrigin+"px"):(!(n=t.style[e])||"auto"===n||i||~(n+"").indexOf("calc("))&&(n=_i[e]&&_i[e](t,e,r)||$r(t,e)||ot(t,e)||("opacity"===e?1:0)),r&&!~(n+"").trim().indexOf(" ")?hi(t,e,n,r)+r:n},li=function(t,e,r,i){if(!r||"none"===r){var n=Kr(e,t,1),s=n&&$r(t,n,1);s&&s!==r?(e=n,r=s):"borderColor"===e&&(r=$r(t,"borderTopColor"))}var a,o,u,h,f,l,c,p,_,d,m,v=new ir(this._pt,t.style,e,0,1,$e),y=0,x=0;if(v.b=r,v.e=i,r+="","auto"===(i+="")&&(t.style[e]=i,i=$r(t,e)||i,t.style[e]=r),me(a=[r,i]),i=a[1],u=(r=a[0]).match(L)||[],(i.match(L)||[]).length){for(;o=L.exec(i);)c=o[0],_=i.substring(y,o.index),f?f=(f+1)%5:"rgba("!==_.substr(-5)&&"hsla("!==_.substr(-5)||(f=1),c!==(l=u[x++]||"")&&(h=parseFloat(l)||0,m=l.substr((h+"").length),"="===c.charAt(1)&&(c=lt(h,c)+m),p=parseFloat(c),d=c.substr((p+"").length),y=L.lastIndex-d.length,d||(d=d||g.units[e]||m,y===i.length&&(i+=d,v.e+=d)),m!==d&&(h=hi(t,e,l,d)||0),v._pt={_next:v._pt,p:_||1===x?_:",",s:h,c:p-h,m:f&&f<4||"zIndex"===e?Math.round:0});v.c=y<i.length?i.substring(y,i.length):""}else v.r="display"===e&&"none"===i?Lr:Br;return Y.test(i)&&(v.e=0),this._pt=v,v},ci={top:"0%",bottom:"100%",left:"0%",right:"100%",center:"50%"},pi=function(t,e){if(e.tween&&e.tween._time===e.tween._dur){var r,i,n,s=e.t,a=s.style,o=e.u,u=s._gsap;if("all"===o||!0===o)a.cssText="",i=1;else for(n=(o=o.split(",")).length;--n>-1;)r=o[n],Mr[r]&&(i=1,r="transformOrigin"===r?jr:Vr),si(s,r);i&&(si(s,Vr),u&&(u.svg&&s.removeAttribute("transform"),Ti(s,1),u.uncache=1,Wr(a)))}},_i={clearProps:function(t,e,r,i,n){if("isFromStart"!==n.data){var s=t._pt=new ir(t._pt,e,r,0,0,pi);return s.u=i,s.pr=-10,s.tween=n,t._props.push(r),1}}},di=[1,0,0,1,0,0],mi={},gi=function(t){return"matrix(1, 0, 0, 1, 0, 0)"===t||"none"===t||!t},vi=function(t){var e=$r(t,Vr);return gi(e)?di:e.substr(7).match(B).map(ht)},yi=function(t,e){var r,i,n,s,a=t._gsap||at(t),o=t.style,u=vi(t);return a.svg&&t.getAttribute("transform")?"1,0,0,1,0,0"===(u=[(n=t.transform.baseVal.consolidate().matrix).a,n.b,n.c,n.d,n.e,n.f]).join(",")?di:u:(u!==di||t.offsetParent||t===vr||a.svg||(n=o.display,o.display="block",(r=t.parentNode)&&t.offsetParent||(s=1,i=t.nextElementSibling,vr.appendChild(t)),u=vi(t),n?o.display=n:si(t,"display"),s&&(i?r.insertBefore(t,i):r?r.appendChild(t):vr.removeChild(t))),e&&u.length>6?[u[0],u[1],u[4],u[5],u[12],u[13]]:u)},xi=function(t,e,r,i,n,s){var a,o,u,h=t._gsap,f=n||yi(t,!0),l=h.xOrigin||0,c=h.yOrigin||0,p=h.xOffset||0,_=h.yOffset||0,d=f[0],m=f[1],g=f[2],v=f[3],y=f[4],x=f[5],T=e.split(" "),w=parseFloat(T[0])||0,b=parseFloat(T[1])||0;r?f!==di&&(o=d*v-m*g)&&(u=w*(-m/o)+b*(d/o)-(d*x-m*y)/o,w=w*(v/o)+b*(-g/o)+(g*x-v*y)/o,b=u):(w=(a=ii(t)).x+(~T[0].indexOf("%")?w/100*a.width:w),b=a.y+(~(T[1]||T[0]).indexOf("%")?b/100*a.height:b)),i||!1!==i&&h.smooth?(y=w-l,x=b-c,h.xOffset=p+(y*d+x*g)-y,h.yOffset=_+(y*m+x*v)-x):h.xOffset=h.yOffset=0,h.xOrigin=w,h.yOrigin=b,h.smooth=!!i,h.origin=e,h.originIsAbsolute=!!r,t.style[jr]="0px 0px",s&&(ai(s,h,"xOrigin",l,w),ai(s,h,"yOrigin",c,b),ai(s,h,"xOffset",p,h.xOffset),ai(s,h,"yOffset",_,h.yOffset)),t.setAttribute("data-svg-origin",w+" "+b)},Ti=function(t,e){var r=t._gsap||new Pe(t);if("x"in r&&!e&&!r.uncache)return r;var i,n,s,a,o,u,h,f,l,c,p,_,d,m,v,y,x,T,w,b,M,O,k,C,A,D,P,S,E,R,z,F,B=t.style,L=r.scaleX<0,I=getComputedStyle(t),Y=$r(t,jr)||"0";return i=n=s=u=h=f=l=c=p=0,a=o=1,r.svg=!(!t.getCTM||!ni(t)),I.translate&&("none"===I.translate&&"none"===I.scale&&"none"===I.rotate||(B[Vr]=("none"!==I.translate?"translate3d("+(I.translate+" 0 0").split(" ").slice(0,3).join(", ")+") ":"")+("none"!==I.rotate?"rotate("+I.rotate+") ":"")+("none"!==I.scale?"scale("+I.scale.split(" ").join(",")+") ":"")+I[Vr]),B.scale=B.rotate=B.translate="none"),m=yi(t,r.svg),r.svg&&(C=(!r.uncache||"0px 0px"===Y)&&!e&&t.getAttribute("data-svg-origin"),xi(t,C||Y,!!C||r.originIsAbsolute,!1!==r.smooth,m)),_=r.xOrigin||0,d=r.yOrigin||0,m!==di&&(T=m[0],w=m[1],b=m[2],M=m[3],i=O=m[4],n=k=m[5],6===m.length?(a=Math.sqrt(T*T+w*w),o=Math.sqrt(M*M+b*b),u=T||w?Cr(w,T)*Or:0,(l=b||M?Cr(b,M)*Or+u:0)&&(o*=Math.abs(Math.cos(l*kr))),r.svg&&(i-=_-(_*T+d*b),n-=d-(_*w+d*M))):(F=m[6],R=m[7],P=m[8],S=m[9],E=m[10],z=m[11],i=m[12],n=m[13],s=m[14],h=(v=Cr(F,E))*Or,v&&(C=O*(y=Math.cos(-v))+P*(x=Math.sin(-v)),A=k*y+S*x,D=F*y+E*x,P=O*-x+P*y,S=k*-x+S*y,E=F*-x+E*y,z=R*-x+z*y,O=C,k=A,F=D),f=(v=Cr(-b,E))*Or,v&&(y=Math.cos(-v),z=M*(x=Math.sin(-v))+z*y,T=C=T*y-P*x,w=A=w*y-S*x,b=D=b*y-E*x),u=(v=Cr(w,T))*Or,v&&(C=T*(y=Math.cos(v))+w*(x=Math.sin(v)),A=O*y+k*x,w=w*y-T*x,k=k*y-O*x,T=C,O=A),h&&Math.abs(h)+Math.abs(u)>359.9&&(h=u=0,f=180-f),a=ht(Math.sqrt(T*T+w*w+b*b)),o=ht(Math.sqrt(k*k+F*F)),v=Cr(O,k),l=Math.abs(v)>2e-4?v*Or:0,p=z?1/(z<0?-z:z):0),r.svg&&(C=t.getAttribute("transform"),r.forceCSS=t.setAttribute("transform","")||!gi($r(t,Vr)),C&&t.setAttribute("transform",C))),Math.abs(l)>90&&Math.abs(l)<270&&(L?(a*=-1,l+=u<=0?180:-180,u+=u<=0?180:-180):(o*=-1,l+=l<=0?180:-180)),e=e||r.uncache,r.x=i-((r.xPercent=i&&(!e&&r.xPercent||(Math.round(t.offsetWidth/2)===Math.round(-i)?-50:0)))?t.offsetWidth*r.xPercent/100:0)+"px",r.y=n-((r.yPercent=n&&(!e&&r.yPercent||(Math.round(t.offsetHeight/2)===Math.round(-n)?-50:0)))?t.offsetHeight*r.yPercent/100:0)+"px",r.z=s+"px",r.scaleX=ht(a),r.scaleY=ht(o),r.rotation=ht(u)+"deg",r.rotationX=ht(h)+"deg",r.rotationY=ht(f)+"deg",r.skewX=l+"deg",r.skewY=c+"deg",r.transformPerspective=p+"px",(r.zOrigin=parseFloat(Y.split(" ")[2])||0)&&(B[jr]=wi(Y)),r.xOffset=r.yOffset=0,r.force3D=g.force3D,r.renderTransform=r.svg?ki:br?Oi:Mi,r.uncache=0,r},wi=function(t){return(t=t.split(" "))[0]+" "+t[1]},bi=function(t,e,r){var i=jt(e);return ht(parseFloat(e)+parseFloat(hi(t,"x",r+"px",i)))+i},Mi=function(t,e){e.z="0px",e.rotationY=e.rotationX="0deg",e.force3D=0,Oi(t,e)},Oi=function(t,e){var r=e||this,i=r.xPercent,n=r.yPercent,s=r.x,a=r.y,o=r.z,u=r.rotation,h=r.rotationY,f=r.rotationX,l=r.skewX,c=r.skewY,p=r.scaleX,_=r.scaleY,d=r.transformPerspective,m=r.force3D,g=r.target,v=r.zOrigin,y="",x="auto"===m&&t&&1!==t||!0===m;if(v&&("0deg"!==f||"0deg"!==h)){var T,w=parseFloat(h)*kr,b=Math.sin(w),M=Math.cos(w);w=parseFloat(f)*kr,T=Math.cos(w),s=bi(g,s,b*T*-v),a=bi(g,a,-Math.sin(w)*-v),o=bi(g,o,M*T*-v+v)}"0px"!==d&&(y+="perspective("+d+") "),(i||n)&&(y+="translate("+i+"%, "+n+"%) "),(x||"0px"!==s||"0px"!==a||"0px"!==o)&&(y+="0px"!==o||x?"translate3d("+s+", "+a+", "+o+") ":"translate("+s+", "+a+") "),"0deg"!==u&&(y+="rotate("+u+") "),"0deg"!==h&&(y+="rotateY("+h+") "),"0deg"!==f&&(y+="rotateX("+f+") "),"0deg"===l&&"0deg"===c||(y+="skew("+l+", "+c+") "),1===p&&1===_||(y+="scale("+p+", "+_+") "),g.style[Vr]=y||"translate(0, 0)"},ki=function(t,e){var r,i,n,s,a,o=e||this,u=o.xPercent,h=o.yPercent,f=o.x,l=o.y,c=o.rotation,p=o.skewX,_=o.skewY,d=o.scaleX,m=o.scaleY,g=o.target,v=o.xOrigin,y=o.yOrigin,x=o.xOffset,T=o.yOffset,w=o.forceCSS,b=parseFloat(f),M=parseFloat(l);c=parseFloat(c),p=parseFloat(p),(_=parseFloat(_))&&(p+=_=parseFloat(_),c+=_),c||p?(c*=kr,p*=kr,r=Math.cos(c)*d,i=Math.sin(c)*d,n=Math.sin(c-p)*-m,s=Math.cos(c-p)*m,p&&(_*=kr,a=Math.tan(p-_),n*=a=Math.sqrt(1+a*a),s*=a,_&&(a=Math.tan(_),r*=a=Math.sqrt(1+a*a),i*=a)),r=ht(r),i=ht(i),n=ht(n),s=ht(s)):(r=d,s=m,i=n=0),(b&&!~(f+"").indexOf("px")||M&&!~(l+"").indexOf("px"))&&(b=hi(g,"x",f,"px"),M=hi(g,"y",l,"px")),(v||y||x||T)&&(b=ht(b+v-(v*r+y*n)+x),M=ht(M+y-(v*i+y*s)+T)),(u||h)&&(a=g.getBBox(),b=ht(b+u/100*a.width),M=ht(M+h/100*a.height)),a="matrix("+r+","+i+","+n+","+s+","+b+","+M+")",g.setAttribute("transform",a),w&&(g.style[Vr]=a)},Ci=function(t,e,r,i,n){var s,a,o=O(n),u=parseFloat(n)*(o&&~n.indexOf("rad")?Or:1)-i,h=i+u+"deg";return o&&("short"===(s=n.split("_")[1])&&(u%=360)!==u%180&&(u+=u<0?360:-360),"cw"===s&&u<0?u=(u+36e9)%360-360*~~(u/360):"ccw"===s&&u>0&&(u=(u-36e9)%360-360*~~(u/360))),t._pt=a=new ir(t._pt,e,r,i,u,Rr),a.e=h,a.u="deg",t._props.push(r),a},Ai=function(t,e){for(var r in e)t[r]=e[r];return t},Di=function(t,e,r){var i,n,s,a,o,u,h,f=Ai({},r._gsap),l=r.style;for(n in f.svg?(s=r.getAttribute("transform"),r.setAttribute("transform",""),l[Vr]=e,i=Ti(r,1),si(r,Vr),r.setAttribute("transform",s)):(s=getComputedStyle(r)[Vr],l[Vr]=e,i=Ti(r,1),l[Vr]=s),Mr)(s=f[n])!==(a=i[n])&&"perspective,force3D,transformOrigin,svgOrigin".indexOf(n)<0&&(o=jt(s)!==(h=jt(a))?hi(r,n,s,h):parseFloat(s),u=parseFloat(a),t._pt=new ir(t._pt,i,n,o,u-o,Er),t._pt.u=h||0,t._props.push(n));Ai(i,f)};ut("padding,margin,Width,Radius",(function(t,e){var r="Top",i="Right",n="Bottom",s="Left",a=(e<3?[r,i,n,s]:[r+s,r+i,n+i,n+s]).map((function(r){return e<2?t+r:"border"+r+t}));_i[e>1?"border"+t:t]=function(t,e,r,i,n){var s,o;if(arguments.length<4)return s=a.map((function(e){return fi(t,e,r)})),5===(o=s.join(" ")).split(s[0]).length?s[0]:o;s=(i+"").split(" "),o={},a.forEach((function(t,e){return o[t]=s[e]=s[e]||s[(e-1)/2|0]})),t.init(e,o,n)}}));var Pi,Si,Ei={name:"css",register:ti,targetTest:function(t){return t.style&&t.nodeType},init:function(t,e,r,i,n){var s,a,o,u,h,f,l,c,p,_,d,m,v,y,x,T,w,b,M,k,C=this._props,A=t.style,D=r.vars.startAt;for(l in yr||ti(),this.styles=this.styles||Hr(t),T=this.styles.props,this.tween=r,e)if("autoRound"!==l&&(a=e[l],!tt[l]||!Le(l,e,r,i,t,n)))if(h=typeof a,f=_i[l],"function"===h&&(h=typeof(a=a.call(r,i,t,n))),"string"===h&&~a.indexOf("random(")&&(a=ie(a)),f)f(this,t,l,a,r)&&(x=1);else if("--"===l.substr(0,2))s=(getComputedStyle(t).getPropertyValue(l)+"").trim(),a+="",_e.lastIndex=0,_e.test(s)||(c=jt(s),p=jt(a)),p?c!==p&&(s=hi(t,l,s,p)+p):c&&(a+=c),this.add(A,"setProperty",s,a,i,n,0,0,l),C.push(l),T.push(l,A[l]);else if("undefined"!==h){if(D&&l in D?(s="function"==typeof D[l]?D[l].call(r,i,t,n):D[l],O(s)&&~s.indexOf("random(")&&(s=ie(s)),jt(s+"")||(s+=g.units[l]||jt(fi(t,l))||""),"="===(s+"").charAt(1)&&(s=fi(t,l))):s=fi(t,l),u=parseFloat(s),(_="string"===h&&"="===a.charAt(1)&&a.substr(0,2))&&(a=a.substr(2)),o=parseFloat(a),l in Sr&&("autoAlpha"===l&&(1===u&&"hidden"===fi(t,"visibility")&&o&&(u=0),T.push("visibility",A.visibility),ai(this,A,"visibility",u?"inherit":"hidden",o?"inherit":"hidden",!o)),"scale"!==l&&"transform"!==l&&~(l=Sr[l]).indexOf(",")&&(l=l.split(",")[0])),d=l in Mr)if(this.styles.save(l),m||((v=t._gsap).renderTransform&&!e.parseTransform||Ti(t,e.parseTransform),y=!1!==e.smoothOrigin&&v.smooth,(m=this._pt=new ir(this._pt,A,Vr,0,1,v.renderTransform,v,0,-1)).dep=1),"scale"===l)this._pt=new ir(this._pt,v,"scaleY",v.scaleY,(_?lt(v.scaleY,_+o):o)-v.scaleY||0,Er),this._pt.u=0,C.push("scaleY",l),l+="X";else{if("transformOrigin"===l){T.push(jr,A[jr]),b=void 0,M=void 0,k=void 0,b=(w=a).split(" "),M=b[0],k=b[1]||"50%","top"!==M&&"bottom"!==M&&"left"!==k&&"right"!==k||(w=M,M=k,k=w),b[0]=ci[M]||M,b[1]=ci[k]||k,a=b.join(" "),v.svg?xi(t,a,0,y,0,this):((p=parseFloat(a.split(" ")[2])||0)!==v.zOrigin&&ai(this,v,"zOrigin",v.zOrigin,p),ai(this,A,l,wi(s),wi(a)));continue}if("svgOrigin"===l){xi(t,a,1,y,0,this);continue}if(l in mi){Ci(this,v,l,u,_?lt(u,_+a):a);continue}if("smoothOrigin"===l){ai(this,v,"smooth",v.smooth,a);continue}if("force3D"===l){v[l]=a;continue}if("transform"===l){Di(this,a,t);continue}}else l in A||(l=Kr(l)||l);if(d||(o||0===o)&&(u||0===u)&&!Pr.test(a)&&l in A)o||(o=0),(c=(s+"").substr((u+"").length))!==(p=jt(a)||(l in g.units?g.units[l]:c))&&(u=hi(t,l,s,p)),this._pt=new ir(this._pt,d?v:A,l,u,(_?lt(u,_+o):o)-u,d||"px"!==p&&"zIndex"!==l||!1===e.autoRound?Er:Fr),this._pt.u=p||0,c!==p&&"%"!==p&&(this._pt.b=s,this._pt.r=zr);else if(l in A)li.call(this,t,l,s,_?_+a:a);else{if(!(l in t)){j(l,a);continue}this.add(t,l,s||t[l],_?_+a:a,i,n)}d||T.push(l,A[l]),C.push(l)}x&&rr(this)},render:function(t,e){if(e.tween._time||!wr())for(var r=e._pt;r;)r.r(t,r.d),r=r._next;else e.styles.revert()},get:fi,aliases:Sr,getSetter:function(t,e,r){var i=Sr[e];return i&&i.indexOf(",")<0&&(e=i),e in Mr&&e!==jr&&(t._gsap.x||fi(t,"x"))?r&&Tr===r?"scale"===e?Xr:Ur:(Tr=r||{})&&("scale"===e?Nr:qr):t.style&&!A(t.style[e])?Ir:~e.indexOf("-")?Yr:Qe(t,e)},core:{_removeProperty:si,_getMatrix:yi}};dr.utils.checkPrefix=Kr,dr.core.getStyleSaver=Hr,Si=ut("x,y,z,scale,scaleX,scaleY,xPercent,yPercent,"+(Pi="rotation,rotationX,rotationY,skewX,skewY")+",transform,transformOrigin,svgOrigin,force3D,smoothOrigin,transformPerspective",(function(t){Mr[t]=1})),ut(Pi,(function(t){g.units[t]="deg",mi[t]=1})),Sr[Si[13]]="x,y,z,scale,scaleX,scaleY,xPercent,yPercent,"+Pi,ut("0:translateX,1:translateY,2:translateZ,8:rotate,8:rotationZ,8:rotateZ,9:rotateX,10:rotateY",(function(t){var e=t.split(":");Sr[e[1]]=Si[e[0]]})),ut("x,y,z,top,right,bottom,left,width,height,fontSize,padding,margin,perspective",(function(t){g.units[t]="px"})),dr.registerPlugin(Ei);var Ri=dr.registerPlugin(Ei)||dr;Ri.core.Tween;var zi=Ri;export{zi as default};
-//# sourceMappingURL=gsap.js.map
