--- public/web_modules/gsap/Draggable.js
@@ -1,11 +1,0 @@
-/*!
- * matrix 3.11.0
- * https://greensock.com
- *
- * Copyright 2008-2022, GreenSock. All rights reserved.
- * Subject to the terms at https://greensock.com/standard-license or for
- * Club GreenSock members, the agreement issued with that membership.
- * @author: Jack Doyle, jack@greensock.com
-*/
-var t,e,n,o,r,i,l,a,s,c="transform",d=c+"Origin",p=function(r){var i=r.ownerDocument||r;!(c in r.style)&&"msTransform"in r.style&&(d=(c="msTransform")+"Origin");for(;i.parentNode&&(i=i.parentNode););if(e=window,l=new v,i){t=i,n=i.documentElement,o=i.body,(a=t.createElementNS("http://www.w3.org/2000/svg","g")).style.transform="none";var p=i.createElement("div"),u=i.createElement("div");o.appendChild(p),p.appendChild(u),p.style.position="static",p.style[c]="translate3d(0,0,1px)",s=u.offsetParent!==p,o.removeChild(p)}return i},u=[],f=[],h=function(t){return t.ownerSVGElement||("svg"===(t.tagName+"").toLowerCase()?t:null)},g=function e(n,o){if(n.parentNode&&(t||p(n))){var l=h(n),a=l?l.getAttribute("xmlns")||"http://www.w3.org/2000/svg":"http://www.w3.org/1999/xhtml",s=l?o?"rect":"g":"div",c=2!==o?0:100,d=3===o?100:0,u="position:absolute;display:block;pointer-events:none;margin:0;padding:0;",f=t.createElementNS?t.createElementNS(a.replace(/^https/,"http"),s):t.createElement(s);return o&&(l?(i||(i=e(n)),f.setAttribute("width",.01),f.setAttribute("height",.01),f.setAttribute("transform","translate("+c+","+d+")"),i.appendChild(f)):(r||((r=e(n)).style.cssText=u),f.style.cssText=u+"width:0.1px;height:0.1px;top:"+d+"px;left:"+c+"px",r.appendChild(f))),f}throw"Need document and parent."},x=function(t,n){var o,p,x,m,y,w,b=h(t),T=t===b,E=b?u:f,M=t.parentNode;if(t===e)return t;if(E.length||E.push(g(t,1),g(t,2),g(t,3)),o=b?i:r,b)T?(m=-(x=function(t){var e,n=t.getCTM();return n||(e=t.style[c],t.style[c]="none",t.appendChild(a),n=a.getCTM(),t.removeChild(a),e?t.style[c]=e:t.style.removeProperty(c.replace(/([A-Z])/g,"-$1").toLowerCase())),n||l.clone()}(t)).e/x.a,y=-x.f/x.d,p=l):t.getBBox?(x=t.getBBox(),m=(p=(p=t.transform?t.transform.baseVal:{}).numberOfItems?p.numberOfItems>1?function(t){for(var e=new v,n=0;n<t.numberOfItems;n++)e.multiply(t.getItem(n).matrix);return e}(p):p.getItem(0).matrix:l).a*x.x+p.c*x.y,y=p.b*x.x+p.d*x.y):(p=new v,m=y=0),n&&"g"===t.tagName.toLowerCase()&&(m=y=0),(T?b:M).appendChild(o),o.setAttribute("transform","matrix("+p.a+","+p.b+","+p.c+","+p.d+","+(p.e+m)+","+(p.f+y)+")");else{if(m=y=0,s)for(p=t.offsetParent,x=t;x&&(x=x.parentNode)&&x!==p&&x.parentNode;)(e.getComputedStyle(x)[c]+"").length>4&&(m=x.offsetLeft,y=x.offsetTop,x=0);if("absolute"!==(w=e.getComputedStyle(t)).position&&"fixed"!==w.position)for(p=t.offsetParent;M&&M!==p;)m+=M.scrollLeft||0,y+=M.scrollTop||0,M=M.parentNode;(x=o.style).top=t.offsetTop-y+"px",x.left=t.offsetLeft-m+"px",x[c]=w[c],x[d]=w[d],x.position="fixed"===w.position?"fixed":"absolute",t.parentNode.appendChild(o)}return o},m=function(t,e,n,o,r,i,l){return t.a=e,t.b=n,t.c=o,t.d=r,t.e=i,t.f=l,t},v=function(){function t(t,e,n,o,r,i){void 0===t&&(t=1),void 0===e&&(e=0),void 0===n&&(n=0),void 0===o&&(o=1),void 0===r&&(r=0),void 0===i&&(i=0),m(this,t,e,n,o,r,i)}var e=t.prototype;return e.inverse=function(){var t=this.a,e=this.b,n=this.c,o=this.d,r=this.e,i=this.f,l=t*o-e*n||1e-10;return m(this,o/l,-e/l,-n/l,t/l,(n*i-o*r)/l,-(t*i-e*r)/l)},e.multiply=function(t){var e=this.a,n=this.b,o=this.c,r=this.d,i=this.e,l=this.f,a=t.a,s=t.c,c=t.b,d=t.d,p=t.e,u=t.f;return m(this,a*e+c*o,a*n+c*r,s*e+d*o,s*n+d*r,i+p*e+u*o,l+p*n+u*r)},e.clone=function(){return new t(this.a,this.b,this.c,this.d,this.e,this.f)},e.equals=function(t){var e=this.a,n=this.b,o=this.c,r=this.d,i=this.e,l=this.f;return e===t.a&&n===t.b&&o===t.c&&r===t.d&&i===t.e&&l===t.f},e.apply=function(t,e){void 0===e&&(e={});var n=t.x,o=t.y,r=this.a,i=this.b,l=this.c,a=this.d,s=this.e,c=this.f;return e.x=n*r+o*l+s||0,e.y=n*i+o*a+c||0,e},t}();function y(r,i,l,a){if(!r||!r.parentNode||(t||p(r)).documentElement===r)return new v;var s=function(t){for(var e,n;t&&t!==o;)(n=t._gsap)&&n.uncache&&n.get(t,"x"),n&&!n.scaleX&&!n.scaleY&&n.renderTransform&&(n.scaleX=n.scaleY=1e-4,n.renderTransform(1,n),e?e.push(n):e=[n]),t=t.parentNode;return e}(r),c=h(r)?u:f,d=x(r,l),g=c[0].getBoundingClientRect(),m=c[1].getBoundingClientRect(),y=c[2].getBoundingClientRect(),w=d.parentNode,b=!a&&function t(n){return"fixed"===e.getComputedStyle(n).position||((n=n.parentNode)&&1===n.nodeType?t(n):void 0)}(r),T=new v((m.left-g.left)/100,(m.top-g.top)/100,(y.left-g.left)/100,(y.top-g.top)/100,g.left+(b?0:e.pageXOffset||t.scrollLeft||n.scrollLeft||o.scrollLeft||0),g.top+(b?0:e.pageYOffset||t.scrollTop||n.scrollTop||o.scrollTop||0));if(w.removeChild(d),s)for(g=s.length;g--;)(m=s[g]).scaleX=m.scaleY=0,m.renderTransform(1,m);return i?T.inverse():T}var w,b,T,E,M,X,S,Y,D,k,C,L,N,P,O,_,R,A,B,I,H,F=function(){return"undefined"!=typeof window},W=function(){return w||F()&&(w=window.gsap)&&w.registerPlugin&&w},z=function(t){return"function"==typeof t},V=function(t){return"object"==typeof t},K=function(t){return void 0===t},G=function(){return!1},U="transform",j="transformOrigin",q=function(t){return Math.round(1e4*t)/1e4},$=Array.isArray,Z=function(t,e){var n=T.createElementNS?T.createElementNS((e||"http://www.w3.org/1999/xhtml").replace(/^https/,"http"),t):T.createElement(t);return n.style?n:T.createElement(t)},J=180/Math.PI,Q=1e20,tt=new v,et=Date.now||function(){return(new Date).getTime()},nt=[],ot={},rt=0,it=/^(?:a|input|textarea|button|select)$/i,lt=0,at={},st={},ct=function(t,e){var n,o={};for(n in t)o[n]=e?t[n]*e:t[n];return o},dt=function t(e,n){for(var o,r=e.length;r--;)n?e[r].style.touchAction=n:e[r].style.removeProperty("touch-action"),(o=e[r].children)&&o.length&&t(o,n)},pt=function(){return nt.forEach((function(t){return t()}))},ut=function(){return!nt.length&&w.ticker.remove(pt)},ft=function(t){for(var e=nt.length;e--;)nt[e]===t&&nt.splice(e,1);w.to(ut,{overwrite:!0,delay:15,duration:0,onComplete:ut,data:"_draggable"})},ht=function(t,e,n,o){if(t.addEventListener){var r=N[e];o=o||(C?{passive:!1}:null),t.addEventListener(r||e,n,o),r&&e!==r&&t.addEventListener(e,n,o)}},gt=function(t,e,n){if(t.removeEventListener){var o=N[e];t.removeEventListener(o||e,n),o&&e!==o&&t.removeEventListener(e,n)}},xt=function(t){t.preventDefault&&t.preventDefault(),t.preventManipulation&&t.preventManipulation()},mt=function t(e){O=e.touches&&P<e.touches.length,gt(e.target,"touchend",t)},vt=function(t){O=t.touches&&P<t.touches.length,ht(t.target,"touchend",mt)},yt=function(t){return b.pageYOffset||t.scrollTop||t.documentElement.scrollTop||t.body.scrollTop||0},wt=function(t){return b.pageXOffset||t.scrollLeft||t.documentElement.scrollLeft||t.body.scrollLeft||0},bt=function t(e,n){ht(e,"scroll",n),Et(e.parentNode)||t(e.parentNode,n)},Tt=function t(e,n){gt(e,"scroll",n),Et(e.parentNode)||t(e.parentNode,n)},Et=function(t){return!(t&&t!==E&&9!==t.nodeType&&t!==T.body&&t!==b&&t.nodeType&&t.parentNode)},Mt=function(t,e){var n="x"===e?"Width":"Height",o="scroll"+n,r="client"+n;return Math.max(0,Et(t)?Math.max(E[o],M[o])-(b["inner"+n]||E[r]||M[r]):t[o]-t[r])},Xt=function t(e,n){var o=Mt(e,"x"),r=Mt(e,"y");Et(e)?e=st:t(e.parentNode,n),e._gsMaxScrollX=o,e._gsMaxScrollY=r,n||(e._gsScrollX=e.scrollLeft||0,e._gsScrollY=e.scrollTop||0)},St=function(t,e,n){var o=t.style;o&&(K(o[e])&&(e=D(e,t)||e),null==n?o.removeProperty&&o.removeProperty(e.replace(/([A-Z])/g,"-$1").toLowerCase()):o[e]=n)},Yt=function(t){return b.getComputedStyle(t instanceof Element?t:t.host||(t.parentNode||{}).host||t)},Dt={},kt=function(t){if(t===b)return Dt.left=Dt.top=0,Dt.width=Dt.right=E.clientWidth||t.innerWidth||M.clientWidth||0,Dt.height=Dt.bottom=(t.innerHeight||0)-20<E.clientHeight?E.clientHeight:t.innerHeight||M.clientHeight||0,Dt;var e=t.ownerDocument||T,n=K(t.pageX)?t.nodeType||K(t.left)||K(t.top)?k(t)[0].getBoundingClientRect():t:{left:t.pageX-wt(e),top:t.pageY-yt(e),right:t.pageX-wt(e)+1,bottom:t.pageY-yt(e)+1};return K(n.right)&&!K(n.width)?(n.right=n.left+n.width,n.bottom=n.top+n.height):K(n.width)&&(n={width:n.right-n.left,height:n.bottom-n.top,right:n.right,left:n.left,bottom:n.bottom,top:n.top}),n},Ct=function(t,e,n){var o,r=t.vars,i=r[n],l=t._listeners[e];return z(i)&&(o=i.apply(r.callbackScope||t,r[n+"Params"]||[t.pointerEvent])),l&&!1===t.dispatchEvent(e)&&(o=!1),o},Lt=function(t,e){var n,o,r,i=k(t)[0];return i.nodeType||i===b?Pt(i,e):K(t.left)?{left:o=t.min||t.minX||t.minRotation||0,top:n=t.min||t.minY||0,width:(t.max||t.maxX||t.maxRotation||0)-o,height:(t.max||t.maxY||0)-n}:(r={x:0,y:0},{left:t.left-r.x,top:t.top-r.y,width:t.width,height:t.height})},Nt={},Pt=function(t,e){e=k(e)[0];var n,o,r,i,l,a,s,c,d,p,u,f,h,g=t.getBBox&&t.ownerSVGElement,x=t.ownerDocument||T;if(t===b)r=yt(x),o=(n=wt(x))+(x.documentElement.clientWidth||t.innerWidth||x.body.clientWidth||0),i=r+((t.innerHeight||0)-20<x.documentElement.clientHeight?x.documentElement.clientHeight:t.innerHeight||x.body.clientHeight||0);else{if(e===b||K(e))return t.getBoundingClientRect();n=r=0,g?(u=(p=t.getBBox()).width,f=p.height):(t.viewBox&&(p=t.viewBox.baseVal)&&(n=p.x||0,r=p.y||0,u=p.width,f=p.height),u||(p="border-box"===(h=Yt(t)).boxSizing,u=(parseFloat(h.width)||t.clientWidth||0)+(p?0:parseFloat(h.borderLeftWidth)+parseFloat(h.borderRightWidth)),f=(parseFloat(h.height)||t.clientHeight||0)+(p?0:parseFloat(h.borderTopWidth)+parseFloat(h.borderBottomWidth)))),o=u,i=f}return t===e?{left:n,top:r,width:o-n,height:i-r}:(a=(l=y(e,!0).multiply(y(t))).apply({x:n,y:r}),s=l.apply({x:o,y:r}),c=l.apply({x:o,y:i}),d=l.apply({x:n,y:i}),{left:n=Math.min(a.x,s.x,c.x,d.x),top:r=Math.min(a.y,s.y,c.y,d.y),width:Math.max(a.x,s.x,c.x,d.x)-n,height:Math.max(a.y,s.y,c.y,d.y)-r})},Ot=function(t,e,n,o,r,i){var l,a,s,c={};if(e)if(1!==r&&e instanceof Array){if(c.end=l=[],s=e.length,V(e[0]))for(a=0;a<s;a++)l[a]=ct(e[a],r);else for(a=0;a<s;a++)l[a]=e[a]*r;n+=1.1,o-=1.1}else z(e)?c.end=function(n){var o,i,l=e.call(t,n);if(1!==r)if(V(l)){for(i in o={},l)o[i]=l[i]*r;l=o}else l*=r;return l}:c.end=e;return(n||0===n)&&(c.max=n),(o||0===o)&&(c.min=o),i&&(c.velocity=0),c},_t=function t(e){var n;return!(!e||!e.getAttribute||e===M)&&(!("true"!==(n=e.getAttribute("data-clickable"))&&("false"===n||!e.onclick&&!it.test(e.nodeName+"")&&"true"!==e.getAttribute("contentEditable")))||t(e.parentNode))},Rt=function(t,e){for(var n,o=t.length;o--;)(n=t[o]).ondragstart=n.onselectstart=e?null:G,w.set(n,{lazy:!0,userSelect:e?"text":"none"})},At=function(t,e){t=w.utils.toArray(t)[0],e=e||{};var n,o,r,i,l,a,s=document.createElement("div"),c=s.style,d=t.firstChild,p=0,u=0,f=t.scrollTop,h=t.scrollLeft,g=t.scrollWidth,x=t.scrollHeight,m=0,v=0,y=0;I&&!1!==e.force3D?(l="translate3d(",a="px,0px)"):U&&(l="translate(",a="px)"),this.scrollTop=function(t,e){if(!arguments.length)return-this.top();this.top(-t,e)},this.scrollLeft=function(t,e){if(!arguments.length)return-this.left();this.left(-t,e)},this.left=function(n,o){if(!arguments.length)return-(t.scrollLeft+u);var r=t.scrollLeft-h,i=u;if((r>2||r<-2)&&!o)return h=t.scrollLeft,w.killTweensOf(this,{left:1,scrollLeft:1}),this.left(-h),void(e.onKill&&e.onKill());(n=-n)<0?(u=n-.5|0,n=0):n>v?(u=n-v|0,n=v):u=0,(u||i)&&(this._skip||(c[U]=l+-u+"px,"+-p+a),u+m>=0&&(c.paddingRight=u+m+"px")),t.scrollLeft=0|n,h=t.scrollLeft},this.top=function(n,o){if(!arguments.length)return-(t.scrollTop+p);var r=t.scrollTop-f,i=p;if((r>2||r<-2)&&!o)return f=t.scrollTop,w.killTweensOf(this,{top:1,scrollTop:1}),this.top(-f),void(e.onKill&&e.onKill());(n=-n)<0?(p=n-.5|0,n=0):n>y?(p=n-y|0,n=y):p=0,(p||i)&&(this._skip||(c[U]=l+-u+"px,"+-p+a)),t.scrollTop=0|n,f=t.scrollTop},this.maxScrollTop=function(){return y},this.maxScrollLeft=function(){return v},this.disable=function(){for(d=s.firstChild;d;)i=d.nextSibling,t.appendChild(d),d=i;t===s.parentNode&&t.removeChild(s)},this.enable=function(){if((d=t.firstChild)!==s){for(;d;)i=d.nextSibling,s.appendChild(d),d=i;t.appendChild(s),this.calibrate()}},this.calibrate=function(e){var i,l,a,d=t.clientWidth===n;f=t.scrollTop,h=t.scrollLeft,d&&t.clientHeight===o&&s.offsetHeight===r&&g===t.scrollWidth&&x===t.scrollHeight&&!e||((p||u)&&(l=this.left(),a=this.top(),this.left(-t.scrollLeft),this.top(-t.scrollTop)),i=Yt(t),d&&!e||(c.display="block",c.width="auto",c.paddingRight="0px",(m=Math.max(0,t.scrollWidth-t.clientWidth))&&(m+=parseFloat(i.paddingLeft)+(H?parseFloat(i.paddingRight):0))),c.display="inline-block",c.position="relative",c.overflow="visible",c.verticalAlign="top",c.boxSizing="content-box",c.width="100%",c.paddingRight=m+"px",H&&(c.paddingBottom=i.paddingBottom),n=t.clientWidth,o=t.clientHeight,g=t.scrollWidth,x=t.scrollHeight,v=t.scrollWidth-n,y=t.scrollHeight-o,r=s.offsetHeight,c.display="block",(l||a)&&(this.left(l),this.top(a)))},this.content=s,this.element=t,this._skip=!1,this.enable()},Bt=function(t){if(F()&&document.body){var e=window&&window.navigator;b=window,T=document,E=T.documentElement,M=T.body,X=Z("div"),B=!!window.PointerEvent,(S=Z("div")).style.cssText="visibility:hidden;height:1px;top:-1px;pointer-events:none;position:relative;clear:both;cursor:grab",A="grab"===S.style.cursor?"grab":"move",_=e&&-1!==e.userAgent.toLowerCase().indexOf("android"),L="ontouchstart"in E&&"orientation"in b||e&&(e.MaxTouchPoints>0||e.msMaxTouchPoints>0),o=Z("div"),r=Z("div"),i=r.style,l=M,i.display="inline-block",i.position="relative",o.style.cssText="width:90px;height:40px;padding:10px;overflow:auto;visibility:hidden",o.appendChild(r),l.appendChild(o),n=r.offsetHeight+18>o.scrollHeight,l.removeChild(o),H=n,N=function(t){for(var e=t.split(","),n=(("onpointerdown"in X?"pointerdown,pointermove,pointerup,pointercancel":"onmspointerdown"in X?"MSPointerDown,MSPointerMove,MSPointerUp,MSPointerCancel":t).split(",")),o={},r=4;--r>-1;)o[e[r]]=n[r],o[n[r]]=e[r];try{E.addEventListener("test",null,Object.defineProperty({},"passive",{get:function(){C=1}}))}catch(t){}return o}("touchstart,touchmove,touchend,touchcancel"),ht(T,"touchcancel",G),ht(b,"touchmove",G),M&&M.addEventListener("touchstart",G),ht(T,"contextmenu",(function(){for(var t in ot)ot[t].isPressed&&ot[t].endDrag()})),w=Y=W()}var n,o,r,i,l;w?(R=w.plugins.inertia,D=w.utils.checkPrefix,U=D(U),j=D(j),k=w.utils.toArray,I=!!D("perspective")):t&&console.warn("Please gsap.registerPlugin(Draggable)")},It=function(t){var e,n;function o(e,n){var r;r=t.call(this)||this,Y||Bt(1),e=k(e)[0],R||(R=w.plugins.inertia),r.vars=n=ct(n||{}),r.target=e,r.x=r.y=r.rotation=0,r.dragResistance=parseFloat(n.dragResistance)||0,r.edgeResistance=isNaN(n.edgeResistance)?1:parseFloat(n.edgeResistance)||0,r.lockAxis=n.lockAxis,r.autoScroll=n.autoScroll||0,r.lockedAxis=null,r.allowEventDefault=!!n.allowEventDefault,w.getProperty(e,"x");var i,l,a,s,c,d,p,u,f,h,g,x,m,M,X,D,C,I,H,F,W,G,U,Z,it,ut,mt,Mt,Dt,Pt,It,Ht,Ft,Wt=(n.type||"x,y").toLowerCase(),zt=~Wt.indexOf("x")||~Wt.indexOf("y"),Vt=-1!==Wt.indexOf("rotation"),Kt=Vt?"rotation":zt?"x":"left",Gt=zt?"y":"top",Ut=!(!~Wt.indexOf("x")&&!~Wt.indexOf("left")&&"scroll"!==Wt),jt=!(!~Wt.indexOf("y")&&!~Wt.indexOf("top")&&"scroll"!==Wt),qt=n.minimumMovement||2,$t=function(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}(r),Zt=k(n.trigger||n.handle||e),Jt={},Qt=0,te=!1,ee=n.autoScrollMarginTop||40,ne=n.autoScrollMarginRight||40,oe=n.autoScrollMarginBottom||40,re=n.autoScrollMarginLeft||40,ie=n.clickableTest||_t,le=0,ae=e._gsap||w.core.getCache(e),se=function t(e){return"fixed"===Yt(e).position||((e=e.parentNode)&&1===e.nodeType?t(e):void 0)}(e),ce=function(t,n){return parseFloat(ae.get(e,t,n))},de=e.ownerDocument||T,pe=function(t){return xt(t),t.stopImmediatePropagation&&t.stopImmediatePropagation(),!1},ue=function t(n){if($t.autoScroll&&$t.isDragging&&(te||C)){var o,r,i,a,s,c,d,p,f=e,h=15*$t.autoScroll;for(te=!1,st.scrollTop=null!=b.pageYOffset?b.pageYOffset:null!=de.documentElement.scrollTop?de.documentElement.scrollTop:de.body.scrollTop,st.scrollLeft=null!=b.pageXOffset?b.pageXOffset:null!=de.documentElement.scrollLeft?de.documentElement.scrollLeft:de.body.scrollLeft,a=$t.pointerX-st.scrollLeft,s=$t.pointerY-st.scrollTop;f&&!r;)o=(r=Et(f.parentNode))?st:f.parentNode,i=r?{bottom:Math.max(E.clientHeight,b.innerHeight||0),right:Math.max(E.clientWidth,b.innerWidth||0),left:0,top:0}:o.getBoundingClientRect(),c=d=0,jt&&((p=o._gsMaxScrollY-o.scrollTop)<0?d=p:s>i.bottom-oe&&p?(te=!0,d=Math.min(p,h*(1-Math.max(0,i.bottom-s)/oe)|0)):s<i.top+ee&&o.scrollTop&&(te=!0,d=-Math.min(o.scrollTop,h*(1-Math.max(0,s-i.top)/ee)|0)),d&&(o.scrollTop+=d)),Ut&&((p=o._gsMaxScrollX-o.scrollLeft)<0?c=p:a>i.right-ne&&p?(te=!0,c=Math.min(p,h*(1-Math.max(0,i.right-a)/ne)|0)):a<i.left+re&&o.scrollLeft&&(te=!0,c=-Math.min(o.scrollLeft,h*(1-Math.max(0,a-i.left)/re)|0)),c&&(o.scrollLeft+=c)),r&&(c||d)&&(b.scrollTo(o.scrollLeft,o.scrollTop),Xe($t.pointerX+c,$t.pointerY+d)),f=o}if(C){var g=$t.x,x=$t.y;Vt?($t.deltaX=g-parseFloat(ae.rotation),$t.rotation=g,ae.rotation=g+"deg",ae.renderTransform(1,ae)):l?(jt&&($t.deltaY=x-l.top(),l.top(x)),Ut&&($t.deltaX=g-l.left(),l.left(g))):zt?(jt&&($t.deltaY=x-parseFloat(ae.y),ae.y=x+"px"),Ut&&($t.deltaX=g-parseFloat(ae.x),ae.x=g+"px"),ae.renderTransform(1,ae)):(jt&&($t.deltaY=x-parseFloat(e.style.top||0),e.style.top=x+"px"),Ut&&($t.deltaX=g-parseFloat(e.style.left||0),e.style.left=g+"px")),!u||n||Mt||(Mt=!0,!1===Ct($t,"drag","onDrag")&&(Ut&&($t.x-=$t.deltaX),jt&&($t.y-=$t.deltaY),t(!0)),Mt=!1)}C=!1},fe=function(t,n){var o,r,i=$t.x,a=$t.y;e._gsap||(ae=w.core.getCache(e)),ae.uncache&&w.getProperty(e,"x"),zt?($t.x=parseFloat(ae.x),$t.y=parseFloat(ae.y)):Vt?$t.x=$t.rotation=parseFloat(ae.rotation):l?($t.y=l.top(),$t.x=l.left()):($t.y=parseFloat(e.style.top||(r=Yt(e))&&r.top)||0,$t.x=parseFloat(e.style.left||(r||{}).left)||0),(H||F||W)&&!n&&($t.isDragging||$t.isThrowing)&&(W&&(at.x=$t.x,at.y=$t.y,(o=W(at)).x!==$t.x&&($t.x=o.x,C=!0),o.y!==$t.y&&($t.y=o.y,C=!0)),H&&(o=H($t.x))!==$t.x&&($t.x=o,Vt&&($t.rotation=o),C=!0),F&&((o=F($t.y))!==$t.y&&($t.y=o),C=!0)),C&&ue(!0),t||($t.deltaX=$t.x-i,$t.deltaY=$t.y-a,Ct($t,"throwupdate","onThrowUpdate"))},he=function(t,e,n,o){return null==e&&(e=-Q),null==n&&(n=Q),z(t)?function(r){var i=$t.isPressed?1-$t.edgeResistance:1;return t.call($t,(r>n?n+(r-n)*i:r<e?e+(r-e)*i:r)*o)*o}:$(t)?function(o){for(var r,i,l=t.length,a=0,s=Q;--l>-1;)(i=(r=t[l])-o)<0&&(i=-i),i<s&&r>=e&&r<=n&&(a=l,s=i);return t[a]}:isNaN(t)?function(t){return t}:function(){return t*o}},ge=function(){var t,o,r,i;p=!1,l?(l.calibrate(),$t.minX=g=-l.maxScrollLeft(),$t.minY=m=-l.maxScrollTop(),$t.maxX=h=$t.maxY=x=0,p=!0):n.bounds&&(t=Lt(n.bounds,e.parentNode),Vt?($t.minX=g=t.left,$t.maxX=h=t.left+t.width,$t.minY=m=$t.maxY=x=0):K(n.bounds.maxX)&&K(n.bounds.maxY)?(o=Lt(e,e.parentNode),$t.minX=g=Math.round(ce(Kt,"px")+t.left-o.left),$t.minY=m=Math.round(ce(Gt,"px")+t.top-o.top),$t.maxX=h=Math.round(g+(t.width-o.width)),$t.maxY=x=Math.round(m+(t.height-o.height))):(t=n.bounds,$t.minX=g=t.minX,$t.minY=m=t.minY,$t.maxX=h=t.maxX,$t.maxY=x=t.maxY),g>h&&($t.minX=h,$t.maxX=h=g,g=$t.minX),m>x&&($t.minY=x,$t.maxY=x=m,m=$t.minY),Vt&&($t.minRotation=g,$t.maxRotation=h),p=!0),n.liveSnap&&(r=!0===n.liveSnap?n.snap||{}:n.liveSnap,i=$(r)||z(r),Vt?(H=he(i?r:r.rotation,g,h,1),F=null):r.points?W=function(t,e,n,o,r,i,l){return i=i&&i<Q?i*i:Q,z(t)?function(a){var s,c,d,p=$t.isPressed?1-$t.edgeResistance:1,u=a.x,f=a.y;return a.x=u=u>n?n+(u-n)*p:u<e?e+(u-e)*p:u,a.y=f=f>r?r+(f-r)*p:f<o?o+(f-o)*p:f,(s=t.call($t,a))!==a&&(a.x=s.x,a.y=s.y),1!==l&&(a.x*=l,a.y*=l),i<Q&&(c=a.x-u)*c+(d=a.y-f)*d>i&&(a.x=u,a.y=f),a}:$(t)?function(e){for(var n,o,r,l,a=t.length,s=0,c=Q;--a>-1;)(l=(n=(r=t[a]).x-e.x)*n+(o=r.y-e.y)*o)<c&&(s=a,c=l);return c<=i?t[s]:e}:function(t){return t}}(i?r:r.points,g,h,m,x,r.radius,l?-1:1):(Ut&&(H=he(i?r:r.x||r.left||r.scrollLeft,g,h,l?-1:1)),jt&&(F=he(i?r:r.y||r.top||r.scrollTop,m,x,l?-1:1))))},xe=function(){$t.isThrowing=!1,Ct($t,"throwcomplete","onThrowComplete")},me=function(){$t.isThrowing=!1},ve=function(t,o){var r,i,a,s;t&&R?(!0===t&&(r=n.snap||n.liveSnap||{},i=$(r)||z(r),t={resistance:(n.throwResistance||n.resistance||1e3)/(Vt?10:1)},Vt?t.rotation=Ot($t,i?r:r.rotation,h,g,1,o):(Ut&&(t[Kt]=Ot($t,i?r:r.points||r.x||r.left,h,g,l?-1:1,o||"x"===$t.lockedAxis)),jt&&(t[Gt]=Ot($t,i?r:r.points||r.y||r.top,x,m,l?-1:1,o||"y"===$t.lockedAxis)),(r.points||$(r)&&V(r[0]))&&(t.linkedProps=Kt+","+Gt,t.radius=r.radius))),$t.isThrowing=!0,s=isNaN(n.overshootTolerance)?1===n.edgeResistance?0:1-$t.edgeResistance+.2:n.overshootTolerance,t.duration||(t.duration={max:Math.max(n.minDuration||0,"maxDuration"in n?n.maxDuration:2),min:isNaN(n.minDuration)?0===s||V(t)&&t.resistance>1e3?0:.5:n.minDuration,overshoot:s}),$t.tween=a=w.to(l||e,{inertia:t,data:"_draggable",onComplete:xe,onInterrupt:me,onUpdate:n.fastMode?Ct:fe,onUpdateParams:n.fastMode?[$t,"onthrowupdate","onThrowUpdate"]:r&&r.radius?[!1,!0]:[]}),n.fastMode||(l&&(l._skip=!0),a.render(1e9,!0,!0),fe(!0,!0),$t.endX=$t.x,$t.endY=$t.y,Vt&&($t.endRotation=$t.x),a.play(0),fe(!0,!0),l&&(l._skip=!1))):p&&$t.applyBounds()},ye=function(t){var n,o=Z;Z=y(e.parentNode,!0),t&&$t.isPressed&&!Z.equals(o||new v)&&(n=o.inverse().apply({x:a,y:s}),Z.apply(n,n),a=n.x,s=n.y),Z.equals(tt)&&(Z=null)},we=function(){var t,n,o,r=1-$t.edgeResistance,i=se?wt(de):0,u=se?yt(de):0;ye(!1),Nt.x=$t.pointerX-i,Nt.y=$t.pointerY-u,Z&&Z.apply(Nt,Nt),a=Nt.x,s=Nt.y,C&&(Xe($t.pointerX,$t.pointerY),ue(!0)),Ht=y(e),l?(ge(),d=l.top(),c=l.left()):(be()?(fe(!0,!0),ge()):$t.applyBounds(),Vt?(t=e.ownerSVGElement?[ae.xOrigin-e.getBBox().x,ae.yOrigin-e.getBBox().y]:(Yt(e)[j]||"0 0").split(" "),D=$t.rotationOrigin=y(e).apply({x:parseFloat(t[0])||0,y:parseFloat(t[1])||0}),fe(!0,!0),n=$t.pointerX-D.x-i,o=D.y-$t.pointerY+u,c=$t.x,d=$t.y=Math.atan2(o,n)*J):(d=ce(Gt,"px"),c=ce(Kt,"px"))),p&&r&&(c>h?c=h+(c-h)/r:c<g&&(c=g-(g-c)/r),Vt||(d>x?d=x+(d-x)/r:d<m&&(d=m-(m-d)/r))),$t.startX=c=q(c),$t.startY=d=q(d)},be=function(){return $t.tween&&$t.tween.isActive()},Te=function(){!S.parentNode||be()||$t.isDragging||S.parentNode.removeChild(S)},Ee=function(t,r){var c;if(!i||$t.isPressed||!t||!("mousedown"!==t.type&&"pointerdown"!==t.type||r)&&et()-le<30&&N[$t.pointerEvent.type])It&&t&&i&&xt(t);else{if(it=be(),Ft=!1,$t.pointerEvent=t,N[t.type]?(U=~t.type.indexOf("touch")?t.currentTarget||t.target:de,ht(U,"touchend",Se),ht(U,"touchmove",Me),ht(U,"touchcancel",Se),ht(de,"touchstart",vt)):(U=null,ht(de,"mousemove",Me)),mt=null,B&&U||(ht(de,"mouseup",Se),t&&t.target&&ht(t.target,"mouseup",Se)),G=ie.call($t,t.target)&&!1===n.dragClickables&&!r)return ht(t.target,"change",Se),Ct($t,"pressInit","onPressInit"),Ct($t,"press","onPress"),Rt(Zt,!0),void(It=!1);var d;if(ut=!(!U||Ut===jt||!1===$t.vars.allowNativeTouchScrolling||$t.vars.allowContextMenu&&t&&(t.ctrlKey||t.which>2))&&(Ut?"y":"x"),(It=!ut&&!$t.allowEventDefault)&&(xt(t),ht(b,"touchforcechange",xt)),t.changedTouches?(t=M=t.changedTouches[0],X=t.identifier):t.pointerId?X=t.pointerId:M=X=null,P++,d=ue,nt.push(d),1===nt.length&&w.ticker.add(pt),s=$t.pointerY=t.pageY,a=$t.pointerX=t.pageX,Ct($t,"pressInit","onPressInit"),(ut||$t.autoScroll)&&Xt(e.parentNode),!e.parentNode||!$t.autoScroll||l||Vt||!e.parentNode._gsMaxScrollX||S.parentNode||e.getBBox||(S.style.width=e.parentNode.scrollWidth+"px",e.parentNode.appendChild(S)),we(),$t.tween&&$t.tween.kill(),$t.isThrowing=!1,w.killTweensOf(l||e,Jt,!0),l&&w.killTweensOf(e,{scrollTo:1},!0),$t.tween=$t.lockedAxis=null,(n.zIndexBoost||!Vt&&!l&&!1!==n.zIndexBoost)&&(e.style.zIndex=o.zIndex++),$t.isPressed=!0,u=!(!n.onDrag&&!$t._listeners.drag),f=!(!n.onMove&&!$t._listeners.move),!1!==n.cursor||n.activeCursor)for(c=Zt.length;--c>-1;)w.set(Zt[c],{cursor:n.activeCursor||n.cursor||("grab"===A?"grabbing":A)});Ct($t,"press","onPress")}},Me=function(t){var n,o,r,l,c,d,p=t;if(i&&!O&&$t.isPressed&&t){if($t.pointerEvent=t,n=t.changedTouches){if((t=n[0])!==M&&t.identifier!==X){for(l=n.length;--l>-1&&(t=n[l]).identifier!==X&&t.target!==e;);if(l<0)return}}else if(t.pointerId&&X&&t.pointerId!==X)return;U&&ut&&!mt&&(Nt.x=t.pageX-(se?wt(de):0),Nt.y=t.pageY-(se?yt(de):0),Z&&Z.apply(Nt,Nt),o=Nt.x,r=Nt.y,((c=Math.abs(o-a))!==(d=Math.abs(r-s))&&(c>qt||d>qt)||_&&ut===mt)&&(mt=c>d&&Ut?"x":"y",ut&&mt!==ut&&ht(b,"touchforcechange",xt),!1!==$t.vars.lockAxisOnTouchScroll&&Ut&&jt&&($t.lockedAxis="x"===mt?"y":"x",z($t.vars.onLockAxis)&&$t.vars.onLockAxis.call($t,p)),_&&ut===mt))?Se(p):($t.allowEventDefault||ut&&(!mt||ut===mt)||!1===p.cancelable?It&&(It=!1):(xt(p),It=!0),$t.autoScroll&&(te=!0),Xe(t.pageX,t.pageY,f))}else It&&t&&i&&xt(t)},Xe=function(t,e,n){var o,r,i,l,u,f,v=1-$t.dragResistance,y=1-$t.edgeResistance,w=$t.pointerX,b=$t.pointerY,T=d,E=$t.x,M=$t.y,X=$t.endX,S=$t.endY,Y=$t.endRotation,k=C;$t.pointerX=t,$t.pointerY=e,se&&(t-=wt(de),e-=yt(de)),Vt?(l=Math.atan2(D.y-e,t-D.x)*J,(u=$t.y-l)>180?(d-=360,$t.y=l):u<-180&&(d+=360,$t.y=l),$t.x!==c||Math.abs(d-l)>qt?($t.y=l,i=c+(d-l)*v):i=c):(Z&&(f=t*Z.a+e*Z.c+Z.e,e=t*Z.b+e*Z.d+Z.f,t=f),(r=e-s)<qt&&r>-qt&&(r=0),(o=t-a)<qt&&o>-qt&&(o=0),($t.lockAxis||$t.lockedAxis)&&(o||r)&&((f=$t.lockedAxis)||($t.lockedAxis=f=Ut&&Math.abs(o)>Math.abs(r)?"y":jt?"x":null,f&&z($t.vars.onLockAxis)&&$t.vars.onLockAxis.call($t,$t.pointerEvent)),"y"===f?r=0:"x"===f&&(o=0)),i=q(c+o*v),l=q(d+r*v)),(H||F||W)&&($t.x!==i||$t.y!==l&&!Vt)&&(W&&(at.x=i,at.y=l,f=W(at),i=q(f.x),l=q(f.y)),H&&(i=q(H(i))),F&&(l=q(F(l)))),p&&(i>h?i=h+Math.round((i-h)*y):i<g&&(i=g+Math.round((i-g)*y)),Vt||(l>x?l=Math.round(x+(l-x)*y):l<m&&(l=Math.round(m+(l-m)*y)))),($t.x!==i||$t.y!==l&&!Vt)&&(Vt?($t.endRotation=$t.x=$t.endX=i,C=!0):(jt&&($t.y=$t.endY=l,C=!0),Ut&&($t.x=$t.endX=i,C=!0)),n&&!1===Ct($t,"move","onMove")?($t.pointerX=w,$t.pointerY=b,d=T,$t.x=E,$t.y=M,$t.endX=X,$t.endY=S,$t.endRotation=Y,C=k):!$t.isDragging&&$t.isPressed&&($t.isDragging=Ft=!0,Ct($t,"dragstart","onDragStart")))},Se=function t(o,r){if(i&&$t.isPressed&&(!o||null==X||r||!(o.pointerId&&o.pointerId!==X&&o.target!==e||o.changedTouches&&!function(t,e){for(var n=t.length;n--;)if(t[n].identifier===e)return!0}(o.changedTouches,X)))){$t.isPressed=!1;var l,a,s,c,d,p=o,u=$t.isDragging,f=$t.vars.allowContextMenu&&o&&(o.ctrlKey||o.which>2),h=w.delayedCall(.001,Te);if(U?(gt(U,"touchend",t),gt(U,"touchmove",Me),gt(U,"touchcancel",t),gt(de,"touchstart",vt)):gt(de,"mousemove",Me),gt(b,"touchforcechange",xt),B&&U||(gt(de,"mouseup",t),o&&o.target&&gt(o.target,"mouseup",t)),C=!1,u&&(Qt=lt=et(),$t.isDragging=!1),ft(ue),G&&!f)return o&&(gt(o.target,"change",t),$t.pointerEvent=p),Rt(Zt,!1),Ct($t,"release","onRelease"),Ct($t,"click","onClick"),void(G=!1);for(a=Zt.length;--a>-1;)St(Zt[a],"cursor",n.cursor||(!1!==n.cursor?A:null));if(P--,o){if((l=o.changedTouches)&&(o=l[0])!==M&&o.identifier!==X){for(a=l.length;--a>-1&&(o=l[a]).identifier!==X&&o.target!==e;);if(a<0&&!r)return}$t.pointerEvent=p,$t.pointerX=o.pageX,$t.pointerY=o.pageY}return f&&p?(xt(p),It=!0,Ct($t,"release","onRelease")):p&&!u?(It=!1,it&&(n.snap||n.bounds)&&ve(n.inertia||n.throwProps),Ct($t,"release","onRelease"),_&&"touchmove"===p.type||-1!==p.type.indexOf("cancel")||(Ct($t,"click","onClick"),et()-le<300&&Ct($t,"doubleclick","onDoubleClick"),c=p.target||e,le=et(),d=function(){le===Dt||!$t.enabled()||$t.isPressed||p.defaultPrevented||(c.click?c.click():de.createEvent&&((s=de.createEvent("MouseEvents")).initMouseEvent("click",!0,!0,b,1,$t.pointerEvent.screenX,$t.pointerEvent.screenY,$t.pointerX,$t.pointerY,!1,!1,!1,!1,0,null),c.dispatchEvent(s)))},_||p.defaultPrevented||w.delayedCall(.05,d))):(ve(n.inertia||n.throwProps),$t.allowEventDefault||!p||!1===n.dragClickables&&ie.call($t,p.target)||!u||ut&&(!mt||ut!==mt)||!1===p.cancelable?It=!1:(It=!0,xt(p)),Ct($t,"release","onRelease")),be()&&h.duration($t.tween.duration()),u&&Ct($t,"dragend","onDragEnd"),!0}It&&o&&i&&xt(o)},Ye=function(t){if(t&&$t.isDragging&&!l){var n=t.target||e.parentNode,o=n.scrollLeft-n._gsScrollX,r=n.scrollTop-n._gsScrollY;(o||r)&&(Z?(a-=o*Z.a+r*Z.c,s-=r*Z.d+o*Z.b):(a-=o,s-=r),n._gsScrollX+=o,n._gsScrollY+=r,Xe($t.pointerX,$t.pointerY))}},De=function(t){var e=et(),n=e-le<100,o=e-Qt<50,r=n&&Dt===le,i=$t.pointerEvent&&$t.pointerEvent.defaultPrevented,l=n&&Pt===le,a=t.isTrusted||null==t.isTrusted&&n&&r;if((r||o&&!1!==$t.vars.suppressClickOnDrag)&&t.stopImmediatePropagation&&t.stopImmediatePropagation(),n&&(!$t.pointerEvent||!$t.pointerEvent.defaultPrevented)&&(!r||a&&!l))return a&&r&&(Pt=le),void(Dt=le);($t.isPressed||o||n)&&(a&&t.detail&&n&&!i||xt(t)),n||o||Ft||(t&&t.target&&($t.pointerEvent=t),Ct($t,"click","onClick"))},ke=function(t){return Z?{x:t.x*Z.a+t.y*Z.c+Z.e,y:t.x*Z.b+t.y*Z.d+Z.f}:{x:t.x,y:t.y}};return(I=o.get(e))&&I.kill(),r.startDrag=function(t,n){var o,r,i,l;Ee(t||$t.pointerEvent,!0),n&&!$t.hitTest(t||$t.pointerEvent)&&(o=kt(t||$t.pointerEvent),r=kt(e),i=ke({x:o.left+o.width/2,y:o.top+o.height/2}),l=ke({x:r.left+r.width/2,y:r.top+r.height/2}),a-=i.x-l.x,s-=i.y-l.y),$t.isDragging||($t.isDragging=Ft=!0,Ct($t,"dragstart","onDragStart"))},r.drag=Me,r.endDrag=function(t){return Se(t||$t.pointerEvent,!0)},r.timeSinceDrag=function(){return $t.isDragging?0:(et()-Qt)/1e3},r.timeSinceClick=function(){return(et()-le)/1e3},r.hitTest=function(t,e){return o.hitTest($t.target,t,e)},r.getDirection=function(t,n){var o,r,i,l,a,s,p="velocity"===t&&R?t:V(t)&&!Vt?"element":"start";return"element"===p&&(a=kt($t.target),s=kt(t)),o="start"===p?$t.x-c:"velocity"===p?R.getVelocity(e,Kt):a.left+a.width/2-(s.left+s.width/2),Vt?o<0?"counter-clockwise":"clockwise":(n=n||2,r="start"===p?$t.y-d:"velocity"===p?R.getVelocity(e,Gt):a.top+a.height/2-(s.top+s.height/2),l=(i=Math.abs(o/r))<1/n?"":o<0?"left":"right",i<n&&(""!==l&&(l+="-"),l+=r<0?"up":"down"),l)},r.applyBounds=function(t,o){var r,i,l,a,s,c;if(t&&n.bounds!==t)return n.bounds=t,$t.update(!0,o);if(fe(!0),ge(),p&&!be()){if(r=$t.x,i=$t.y,r>h?r=h:r<g&&(r=g),i>x?i=x:i<m&&(i=m),($t.x!==r||$t.y!==i)&&(l=!0,$t.x=$t.endX=r,Vt?$t.endRotation=r:$t.y=$t.endY=i,C=!0,ue(!0),$t.autoScroll&&!$t.isDragging))for(Xt(e.parentNode),a=e,st.scrollTop=null!=b.pageYOffset?b.pageYOffset:null!=de.documentElement.scrollTop?de.documentElement.scrollTop:de.body.scrollTop,st.scrollLeft=null!=b.pageXOffset?b.pageXOffset:null!=de.documentElement.scrollLeft?de.documentElement.scrollLeft:de.body.scrollLeft;a&&!c;)s=(c=Et(a.parentNode))?st:a.parentNode,jt&&s.scrollTop>s._gsMaxScrollY&&(s.scrollTop=s._gsMaxScrollY),Ut&&s.scrollLeft>s._gsMaxScrollX&&(s.scrollLeft=s._gsMaxScrollX),a=s;$t.isThrowing&&(l||$t.endX>h||$t.endX<g||$t.endY>x||$t.endY<m)&&ve(n.inertia||n.throwProps,l)}return $t},r.update=function(t,n,o){if(n&&$t.isPressed){var r=y(e),i=Ht.apply({x:$t.x-c,y:$t.y-d}),l=y(e.parentNode,!0);l.apply({x:r.e-i.x,y:r.f-i.y},i),$t.x-=i.x-l.e,$t.y-=i.y-l.f,ue(!0),we()}var a=$t.x,s=$t.y;return ye(!n),t?$t.applyBounds():(C&&o&&ue(!0),fe(!0)),n&&(Xe($t.pointerX,$t.pointerY),C&&ue(!0)),$t.isPressed&&!n&&(Ut&&Math.abs(a-$t.x)>.01||jt&&Math.abs(s-$t.y)>.01&&!Vt)&&we(),$t.autoScroll&&(Xt(e.parentNode,$t.isDragging),te=$t.isDragging,ue(!0),Tt(e,Ye),bt(e,Ye)),$t},r.enable=function(t){var o,r,a,s={lazy:!0};if(!1!==n.cursor&&(s.cursor=n.cursor||A),w.utils.checkPrefix("touchCallout")&&(s.touchCallout="none"),"soft"!==t){for(dt(Zt,Ut===jt?"none":n.allowNativeTouchScrolling&&e.scrollHeight===e.clientHeight==(e.scrollWidth===e.clientHeight)||n.allowEventDefault?"manipulation":Ut?"pan-y":"pan-x"),r=Zt.length;--r>-1;)a=Zt[r],B||ht(a,"mousedown",Ee),ht(a,"touchstart",Ee),ht(a,"click",De,!0),w.set(a,s),a.getBBox&&a.ownerSVGElement&&w.set(a.ownerSVGElement,{touchAction:Ut===jt?"none":n.allowNativeTouchScrolling||n.allowEventDefault?"manipulation":Ut?"pan-y":"pan-x"}),n.allowContextMenu||ht(a,"contextmenu",pe);Rt(Zt,!1)}return bt(e,Ye),i=!0,R&&"soft"!==t&&R.track(l||e,zt?"x,y":Vt?"rotation":"top,left"),e._gsDragID=o="d"+rt++,ot[o]=$t,l&&(l.enable(),l.element._gsDragID=o),(n.bounds||Vt)&&we(),n.bounds&&$t.applyBounds(),$t},r.disable=function(t){for(var n,o=$t.isDragging,r=Zt.length;--r>-1;)St(Zt[r],"cursor",null);if("soft"!==t){for(dt(Zt,null),r=Zt.length;--r>-1;)n=Zt[r],St(n,"touchCallout",null),gt(n,"mousedown",Ee),gt(n,"touchstart",Ee),gt(n,"click",De),gt(n,"contextmenu",pe);Rt(Zt,!0),U&&(gt(U,"touchcancel",Se),gt(U,"touchend",Se),gt(U,"touchmove",Me)),gt(de,"mouseup",Se),gt(de,"mousemove",Me)}return Tt(e,Ye),i=!1,R&&"soft"!==t&&R.untrack(l||e,zt?"x,y":Vt?"rotation":"top,left"),l&&l.disable(),ft(ue),$t.isDragging=$t.isPressed=G=!1,o&&Ct($t,"dragend","onDragEnd"),$t},r.enabled=function(t,e){return arguments.length?t?$t.enable(e):$t.disable(e):i},r.kill=function(){return $t.isThrowing=!1,$t.tween&&$t.tween.kill(),$t.disable(),w.set(Zt,{clearProps:"userSelect"}),delete ot[e._gsDragID],$t},~Wt.indexOf("scroll")&&(l=r.scrollProxy=new At(e,function(t,e){for(var n in e)n in t||(t[n]=e[n]);return t}({onKill:function(){$t.isPressed&&Se(null)}},n)),e.style.overflowY=jt&&!L?"auto":"hidden",e.style.overflowX=Ut&&!L?"auto":"hidden",e=l.content),Vt?Jt.rotation=1:(Ut&&(Jt[Kt]=1),jt&&(Jt[Gt]=1)),ae.force3D=!("force3D"in n)||n.force3D,r.enable(),r}return n=t,(e=o).prototype=Object.create(n.prototype),e.prototype.constructor=e,e.__proto__=n,o.register=function(t){w=t,Bt()},o.create=function(t,e){return Y||Bt(!0),k(t).map((function(t){return new o(t,e)}))},o.get=function(t){return ot[(k(t)[0]||{})._gsDragID]},o.timeSinceDrag=function(){return(et()-lt)/1e3},o.hitTest=function(t,e,n){if(t===e)return!1;var o,r,i,l=kt(t),a=kt(e),s=l.top,c=l.left,d=l.right,p=l.bottom,u=l.width,f=l.height,h=a.left>d||a.right<c||a.top>p||a.bottom<s;return h||!n?!h:(i=-1!==(n+"").indexOf("%"),n=parseFloat(n)||0,(o={left:Math.max(c,a.left),top:Math.max(s,a.top)}).width=Math.min(d,a.right)-o.left,o.height=Math.min(p,a.bottom)-o.top,!(o.width<0||o.height<0)&&(i?(n*=.01,(r=o.width*o.height)>=u*f*n||r>=a.width*a.height*n):o.width>n&&o.height>n))},o}(function(){function t(t){this._listeners={},this.target=t||this}var e=t.prototype;return e.addEventListener=function(t,e){var n=this._listeners[t]||(this._listeners[t]=[]);~n.indexOf(e)||n.push(e)},e.removeEventListener=function(t,e){var n=this._listeners[t],o=n&&n.indexOf(e);o>=0&&n.splice(o,1)},e.dispatchEvent=function(t){var e,n=this;return(this._listeners[t]||[]).forEach((function(o){return!1===o.call(n,{type:t,target:n.target})&&(e=!1)})),e},t}());!function(t,e){for(var n in e)n in t||(t[n]=e[n])}(It.prototype,{pointerX:0,pointerY:0,startX:0,startY:0,deltaX:0,deltaY:0,isDragging:!1,isPressed:!1}),It.zIndex=1e3,It.version="3.11.0",W()&&w.registerPlugin(It);export{It as Draggable,It as default};
-//# sourceMappingURL=Draggable.js.map
