--- public/web_modules/immer.js
@@ -1,2 +1,0 @@
-function t(t){for(var e=arguments.length,r=Array(e>1?e-1:0),n=1;n<e;n++)r[n-1]=arguments[n];throw Error("[Immer] minified error nr: "+t+(r.length?" "+r.map((function(t){return"'"+t+"'"})).join(","):"")+". Find the full error at: https://bit.ly/3cXEKWf")}function e(t){return!!t&&!!t[Q]}function r(t){return!!t&&(function(t){if(!t||"object"!=typeof t)return!1;var e=Object.getPrototypeOf(t);if(null===e)return!0;var r=Object.hasOwnProperty.call(e,"constructor")&&e.constructor;return r===Object||"function"==typeof r&&Function.toString.call(r)===Y}(t)||Array.isArray(t)||!!t[L]||!!t.constructor[L]||s(t)||l(t))}function n(r){return e(r)||t(23,r),r[Q].t}function o(t,e,r){void 0===r&&(r=!1),0===i(t)?(r?Object.keys:Z)(t).forEach((function(n){r&&"symbol"==typeof n||e(n,t[n],t)})):t.forEach((function(r,n){return e(n,r,t)}))}function i(t){var e=t[Q];return e?e.i>3?e.i-4:e.i:Array.isArray(t)?1:s(t)?2:l(t)?3:0}function u(t,e){return 2===i(t)?t.has(e):Object.prototype.hasOwnProperty.call(t,e)}function a(t,e){return 2===i(t)?t.get(e):t[e]}function c(t,e,r){var n=i(t);2===n?t.set(e,r):3===n?(t.delete(e),t.add(r)):t[e]=r}function f(t,e){return t===e?0!==t||1/t==1/e:t!=t&&e!=e}function s(t){return q&&t instanceof Map}function l(t){return B&&t instanceof Set}function h(t){return t.o||t.t}function p(t){if(Array.isArray(t))return Array.prototype.slice.call(t);var e=tt(t);delete e[Q];for(var r=Z(e),n=0;n<r.length;n++){var o=r[n],i=e[o];!1===i.writable&&(i.writable=!0,i.configurable=!0),(i.get||i.set)&&(e[o]={configurable:!0,writable:!0,enumerable:i.enumerable,value:t[o]})}return Object.create(Object.getPrototypeOf(t),e)}function v(t,n){return void 0===n&&(n=!1),d(t)||e(t)||!r(t)||(i(t)>1&&(t.set=t.add=t.clear=t.delete=y),Object.freeze(t),n&&o(t,(function(t,e){return v(e,!0)}),!0)),t}function y(){t(2)}function d(t){return null==t||"object"!=typeof t||Object.isFrozen(t)}function g(e){var r=et[e];return r||t(18,e),r}function b(t,e){et[t]||(et[t]=e)}function P(){return T}function m(t,e){e&&(g("Patches"),t.u=[],t.s=[],t.v=e)}function O(t){w(t),t.p.forEach(j),t.p=null}function w(t){t===T&&(T=t.l)}function A(t){return T={p:[],l:T,h:t,m:!0,_:0}}function j(t){var e=t[Q];0===e.i||1===e.i?e.j():e.O=!0}function D(e,n){n._=n.p.length;var o=n.p[0],i=void 0!==e&&e!==o;return n.h.g||g("ES5").S(n,e,i),i?(o[Q].P&&(O(n),t(4)),r(e)&&(e=S(n,e),n.l||k(n,e)),n.u&&g("Patches").M(o[Q].t,e,n.u,n.s)):e=S(n,o,[]),O(n),n.u&&n.v(n.u,n.s),e!==H?e:void 0}function S(t,e,r){if(d(e))return e;var n=e[Q];if(!n)return o(e,(function(o,i){return _(t,n,e,o,i,r)}),!0),e;if(n.A!==t)return e;if(!n.P)return k(t,n.t,!0),n.t;if(!n.I){n.I=!0,n.A._--;var i=4===n.i||5===n.i?n.o=p(n.k):n.o;o(3===n.i?new Set(i):i,(function(e,o){return _(t,n,i,e,o,r)})),k(t,i,!1),r&&t.u&&g("Patches").R(n,r,t.u,t.s)}return n.o}function _(t,n,o,i,a,f){if(e(a)){var s=S(t,a,f&&n&&3!==n.i&&!u(n.D,i)?f.concat(i):void 0);if(c(o,i,s),!e(s))return;t.m=!1}if(r(a)&&!d(a)){if(!t.h.F&&t._<1)return;S(t,a),n&&n.A.l||k(t,a)}}function k(t,e,r){void 0===r&&(r=!1),t.h.F&&t.m&&v(e,r)}function x(t,e){var r=t[Q];return(r?h(r):t)[e]}function E(t,e){if(e in t)for(var r=Object.getPrototypeOf(t);r;){var n=Object.getOwnPropertyDescriptor(r,e);if(n)return n;r=Object.getPrototypeOf(r)}}function M(t){t.P||(t.P=!0,t.l&&M(t.l))}function z(t){t.o||(t.o=p(t.t))}function F(t,e,r){var n=s(e)?g("MapSet").N(e,r):l(e)?g("MapSet").T(e,r):t.g?function(t,e){var r=Array.isArray(t),n={i:r?1:0,A:e?e.A:P(),P:!1,I:!1,D:{},l:e,t:t,k:null,o:null,j:null,C:!1},o=n,i=rt;r&&(o=[n],i=nt);var u=Proxy.revocable(o,i),a=u.revoke,c=u.proxy;return n.k=c,n.j=a,c}(e,r):g("ES5").J(e,r);return(r?r.A:P()).p.push(n),n}function I(n){return e(n)||t(22,n),function t(e){if(!r(e))return e;var n,u=e[Q],f=i(e);if(u){if(!u.P&&(u.i<4||!g("ES5").K(u)))return u.t;u.I=!0,n=R(e,f),u.I=!1}else n=R(e,f);return o(n,(function(e,r){u&&a(u.t,e)===r||c(n,e,t(r))})),3===f?new Set(n):n}(n)}function R(t,e){switch(e){case 2:return new Map(t);case 3:return Array.from(t)}return p(t)}function K(){function t(t,e){var r=a[t];return r?r.enumerable=e:a[t]=r={configurable:!0,enumerable:e,get:function(){var e=this[Q];return rt.get(e,t)},set:function(e){var r=this[Q];rt.set(r,t,e)}},r}function r(t){for(var e=t.length-1;e>=0;e--){var r=t[e][Q];if(!r.P)switch(r.i){case 5:i(r)&&M(r);break;case 4:n(r)&&M(r)}}}function n(t){for(var e=t.t,r=t.k,n=Z(r),o=n.length-1;o>=0;o--){var i=n[o];if(i!==Q){var a=e[i];if(void 0===a&&!u(e,i))return!0;var c=r[i],s=c&&c[Q];if(s?s.t!==a:!f(c,a))return!0}}var l=!!e[Q];return n.length!==Z(e).length+(l?0:1)}function i(t){var e=t.k;if(e.length!==t.t.length)return!0;var r=Object.getOwnPropertyDescriptor(e,e.length-1);if(r&&!r.get)return!0;for(var n=0;n<e.length;n++)if(!e.hasOwnProperty(n))return!0;return!1}var a={};b("ES5",{J:function(e,r){var n=Array.isArray(e),o=function(e,r){if(e){for(var n=Array(r.length),o=0;o<r.length;o++)Object.defineProperty(n,""+o,t(o,!0));return n}var i=tt(r);delete i[Q];for(var u=Z(i),a=0;a<u.length;a++){var c=u[a];i[c]=t(c,e||!!i[c].enumerable)}return Object.create(Object.getPrototypeOf(r),i)}(n,e),i={i:n?5:4,A:r?r.A:P(),P:!1,I:!1,D:{},l:r,t:e,k:o,o:null,O:!1,C:!1};return Object.defineProperty(o,Q,{value:i,writable:!0}),o},S:function(t,n,a){a?e(n)&&n[Q].A===t&&r(t.p):(t.u&&function t(e){if(e&&"object"==typeof e){var r=e[Q];if(r){var n=r.t,a=r.k,c=r.D,f=r.i;if(4===f)o(a,(function(e){e!==Q&&(void 0!==n[e]||u(n,e)?c[e]||t(a[e]):(c[e]=!0,M(r)))})),o(n,(function(t){void 0!==a[t]||u(a,t)||(c[t]=!1,M(r))}));else if(5===f){if(i(r)&&(M(r),c.length=!0),a.length<n.length)for(var s=a.length;s<n.length;s++)c[s]=!1;else for(var l=n.length;l<a.length;l++)c[l]=!0;for(var h=Math.min(a.length,n.length),p=0;p<h;p++)a.hasOwnProperty(p)||(c[p]=!0),void 0===c[p]&&t(a[p])}}}}(t.p[0]),r(t.p))},K:function(t){return 4===t.i?n(t):i(t)}})}function C(){function n(t){if(!r(t))return t;if(Array.isArray(t))return t.map(n);if(s(t))return new Map(Array.from(t.entries()).map((function(t){return[t[0],n(t[1])]})));if(l(t))return new Set(Array.from(t).map(n));var e=Object.create(Object.getPrototypeOf(t));for(var o in t)e[o]=n(t[o]);return u(t,L)&&(e[L]=t[L]),e}function c(t){return e(t)?n(t):t}var f="add";b("Patches",{$:function(e,r){return r.forEach((function(r){for(var o=r.path,u=r.op,c=e,s=0;s<o.length-1;s++){var l=i(c),h=""+o[s];0!==l&&1!==l||"__proto__"!==h&&"constructor"!==h||t(24),"function"==typeof c&&"prototype"===h&&t(24),"object"!=typeof(c=a(c,h))&&t(15,o.join("/"))}var p=i(c),v=n(r.value),y=o[o.length-1];switch(u){case"replace":switch(p){case 2:return c.set(y,v);case 3:t(16);default:return c[y]=v}case f:switch(p){case 1:return"-"===y?c.push(v):c.splice(y,0,v);case 2:return c.set(y,v);case 3:return c.add(v);default:return c[y]=v}case"remove":switch(p){case 1:return c.splice(y,1);case 2:return c.delete(y);case 3:return c.delete(r.value);default:return delete c[y]}default:t(17,u)}})),e},R:function(t,e,r,n){switch(t.i){case 0:case 4:case 2:return function(t,e,r,n){var i=t.t,s=t.o;o(t.D,(function(t,o){var l=a(i,t),h=a(s,t),p=o?u(i,t)?"replace":f:"remove";if(l!==h||"replace"!==p){var v=e.concat(t);r.push("remove"===p?{op:p,path:v}:{op:p,path:v,value:h}),n.push(p===f?{op:"remove",path:v}:"remove"===p?{op:f,path:v,value:c(l)}:{op:"replace",path:v,value:c(l)})}}))}(t,e,r,n);case 5:case 1:return function(t,e,r,n){var o=t.t,i=t.D,u=t.o;if(u.length<o.length){var a=[u,o];o=a[0],u=a[1];var s=[n,r];r=s[0],n=s[1]}for(var l=0;l<o.length;l++)if(i[l]&&u[l]!==o[l]){var h=e.concat([l]);r.push({op:"replace",path:h,value:c(u[l])}),n.push({op:"replace",path:h,value:c(o[l])})}for(var p=o.length;p<u.length;p++){var v=e.concat([p]);r.push({op:f,path:v,value:c(u[p])})}o.length<u.length&&n.push({op:"replace",path:e.concat(["length"]),value:o.length})}(t,e,r,n);case 3:return function(t,e,r,n){var o=t.t,i=t.o,u=0;o.forEach((function(t){if(!i.has(t)){var o=e.concat([u]);r.push({op:"remove",path:o,value:t}),n.unshift({op:f,path:o,value:t})}u++})),u=0,i.forEach((function(t){if(!o.has(t)){var i=e.concat([u]);r.push({op:f,path:i,value:t}),n.unshift({op:"remove",path:i,value:t})}u++}))}(t,e,r,n)}},M:function(t,e,r,n){r.push({op:"replace",path:[],value:e===H?void 0:e}),n.push({op:"replace",path:[],value:t})}})}function N(){function e(t,e){function r(){this.constructor=t}a(t,e),t.prototype=(r.prototype=e.prototype,new r)}function n(t){t.o||(t.D=new Map,t.o=new Map(t.t))}function i(t){t.o||(t.o=new Set,t.t.forEach((function(e){if(r(e)){var n=F(t.A.h,e,t);t.p.set(e,n),t.o.add(n)}else t.o.add(e)})))}function u(e){e.O&&t(3,JSON.stringify(h(e)))}var a=function(t,e){return(a=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(t,e){t.__proto__=e}||function(t,e){for(var r in e)e.hasOwnProperty(r)&&(t[r]=e[r])})(t,e)},c=function(){function t(t,e){return this[Q]={i:2,l:e,A:e?e.A:P(),P:!1,I:!1,o:void 0,D:void 0,t:t,k:this,C:!1,O:!1},this}e(t,Map);var i=t.prototype;return Object.defineProperty(i,"size",{get:function(){return h(this[Q]).size}}),i.has=function(t){return h(this[Q]).has(t)},i.set=function(t,e){var r=this[Q];return u(r),h(r).has(t)&&h(r).get(t)===e||(n(r),M(r),r.D.set(t,!0),r.o.set(t,e),r.D.set(t,!0)),this},i.delete=function(t){if(!this.has(t))return!1;var e=this[Q];return u(e),n(e),M(e),e.t.has(t)?e.D.set(t,!1):e.D.delete(t),e.o.delete(t),!0},i.clear=function(){var t=this[Q];u(t),h(t).size&&(n(t),M(t),t.D=new Map,o(t.t,(function(e){t.D.set(e,!1)})),t.o.clear())},i.forEach=function(t,e){var r=this;h(this[Q]).forEach((function(n,o){t.call(e,r.get(o),o,r)}))},i.get=function(t){var e=this[Q];u(e);var o=h(e).get(t);if(e.I||!r(o))return o;if(o!==e.t.get(t))return o;var i=F(e.A.h,o,e);return n(e),e.o.set(t,i),i},i.keys=function(){return h(this[Q]).keys()},i.values=function(){var t,e=this,r=this.keys();return(t={})[V]=function(){return e.values()},t.next=function(){var t=r.next();return t.done?t:{done:!1,value:e.get(t.value)}},t},i.entries=function(){var t,e=this,r=this.keys();return(t={})[V]=function(){return e.entries()},t.next=function(){var t=r.next();if(t.done)return t;var n=e.get(t.value);return{done:!1,value:[t.value,n]}},t},i[V]=function(){return this.entries()},t}(),f=function(){function t(t,e){return this[Q]={i:3,l:e,A:e?e.A:P(),P:!1,I:!1,o:void 0,t:t,k:this,p:new Map,O:!1,C:!1},this}e(t,Set);var r=t.prototype;return Object.defineProperty(r,"size",{get:function(){return h(this[Q]).size}}),r.has=function(t){var e=this[Q];return u(e),e.o?!!e.o.has(t)||!(!e.p.has(t)||!e.o.has(e.p.get(t))):e.t.has(t)},r.add=function(t){var e=this[Q];return u(e),this.has(t)||(i(e),M(e),e.o.add(t)),this},r.delete=function(t){if(!this.has(t))return!1;var e=this[Q];return u(e),i(e),M(e),e.o.delete(t)||!!e.p.has(t)&&e.o.delete(e.p.get(t))},r.clear=function(){var t=this[Q];u(t),h(t).size&&(i(t),M(t),t.o.clear())},r.values=function(){var t=this[Q];return u(t),i(t),t.o.values()},r.entries=function(){var t=this[Q];return u(t),i(t),t.o.entries()},r.keys=function(){return this.values()},r[V]=function(){return this.values()},r.forEach=function(t,e){for(var r=this.values(),n=r.next();!n.done;)t.call(e,n.value,n.value,this),n=r.next()},t}();b("MapSet",{N:function(t,e){return new c(t,e)},T:function(t,e){return new f(t,e)}})}function W(){K(),N(),C()}function U(t){return t}function $(t){return t}var J,T,X="undefined"!=typeof Symbol&&"symbol"==typeof Symbol("x"),q="undefined"!=typeof Map,B="undefined"!=typeof Set,G="undefined"!=typeof Proxy&&void 0!==Proxy.revocable&&"undefined"!=typeof Reflect,H=X?Symbol.for("immer-nothing"):((J={})["immer-nothing"]=!0,J),L=X?Symbol.for("immer-draftable"):"__$immer_draftable",Q=X?Symbol.for("immer-state"):"__$immer_state",V="undefined"!=typeof Symbol&&Symbol.iterator||"@@iterator",Y=""+Object.prototype.constructor,Z="undefined"!=typeof Reflect&&Reflect.ownKeys?Reflect.ownKeys:void 0!==Object.getOwnPropertySymbols?function(t){return Object.getOwnPropertyNames(t).concat(Object.getOwnPropertySymbols(t))}:Object.getOwnPropertyNames,tt=Object.getOwnPropertyDescriptors||function(t){var e={};return Z(t).forEach((function(r){e[r]=Object.getOwnPropertyDescriptor(t,r)})),e},et={},rt={get:function(t,e){if(e===Q)return t;var n=h(t);if(!u(n,e))return function(t,e,r){var n,o=E(e,r);return o?"value"in o?o.value:null===(n=o.get)||void 0===n?void 0:n.call(t.k):void 0}(t,n,e);var o=n[e];return t.I||!r(o)?o:o===x(t.t,e)?(z(t),t.o[e]=F(t.A.h,o,t)):o},has:function(t,e){return e in h(t)},ownKeys:function(t){return Reflect.ownKeys(h(t))},set:function(t,e,r){var n=E(h(t),e);if(null==n?void 0:n.set)return n.set.call(t.k,r),!0;if(!t.P){var o=x(h(t),e),i=null==o?void 0:o[Q];if(i&&i.t===r)return t.o[e]=r,t.D[e]=!1,!0;if(f(r,o)&&(void 0!==r||u(t.t,e)))return!0;z(t),M(t)}return t.o[e]===r&&"number"!=typeof r&&(void 0!==r||e in t.o)||(t.o[e]=r,t.D[e]=!0,!0)},deleteProperty:function(t,e){return void 0!==x(t.t,e)||e in t.t?(t.D[e]=!1,z(t),M(t)):delete t.D[e],t.o&&delete t.o[e],!0},getOwnPropertyDescriptor:function(t,e){var r=h(t),n=Reflect.getOwnPropertyDescriptor(r,e);return n?{writable:!0,configurable:1!==t.i||"length"!==e,enumerable:n.enumerable,value:r[e]}:n},defineProperty:function(){t(11)},getPrototypeOf:function(t){return Object.getPrototypeOf(t.t)},setPrototypeOf:function(){t(12)}},nt={};o(rt,(function(t,e){nt[t]=function(){return arguments[0]=arguments[0][0],e.apply(this,arguments)}})),nt.deleteProperty=function(t,e){return nt.set.call(this,t,e,void 0)},nt.set=function(t,e,r){return rt.set.call(this,t[0],e,r,t[0])};var ot=function(){function n(e){var n=this;this.g=G,this.F=!0,this.produce=function(e,o,i){if("function"==typeof e&&"function"!=typeof o){var u=o;o=e;var a=n;return function(t){var e=this;void 0===t&&(t=u);for(var r=arguments.length,n=Array(r>1?r-1:0),i=1;i<r;i++)n[i-1]=arguments[i];return a.produce(t,(function(t){var r;return(r=o).call.apply(r,[e,t].concat(n))}))}}var c;if("function"!=typeof o&&t(6),void 0!==i&&"function"!=typeof i&&t(7),r(e)){var f=A(n),s=F(n,e,void 0),l=!0;try{c=o(s),l=!1}finally{l?O(f):w(f)}return"undefined"!=typeof Promise&&c instanceof Promise?c.then((function(t){return m(f,i),D(t,f)}),(function(t){throw O(f),t})):(m(f,i),D(c,f))}if(!e||"object"!=typeof e){if(void 0===(c=o(e))&&(c=e),c===H&&(c=void 0),n.F&&v(c,!0),i){var h=[],p=[];g("Patches").M(e,c,h,p),i(h,p)}return c}t(21,e)},this.produceWithPatches=function(t,e){if("function"==typeof t)return function(e){for(var r=arguments.length,o=Array(r>1?r-1:0),i=1;i<r;i++)o[i-1]=arguments[i];return n.produceWithPatches(e,(function(e){return t.apply(void 0,[e].concat(o))}))};var r,o,i=n.produce(t,e,(function(t,e){r=t,o=e}));return"undefined"!=typeof Promise&&i instanceof Promise?i.then((function(t){return[t,r,o]})):[i,r,o]},"boolean"==typeof(null==e?void 0:e.useProxies)&&this.setUseProxies(e.useProxies),"boolean"==typeof(null==e?void 0:e.autoFreeze)&&this.setAutoFreeze(e.autoFreeze)}var o=n.prototype;return o.createDraft=function(n){r(n)||t(8),e(n)&&(n=I(n));var o=A(this),i=F(this,n,void 0);return i[Q].C=!0,w(o),i},o.finishDraft=function(t,e){var r=(t&&t[Q]).A;return m(r,e),D(void 0,r)},o.setAutoFreeze=function(t){this.F=t},o.setUseProxies=function(e){e&&!G&&t(20),this.g=e},o.applyPatches=function(t,r){var n;for(n=r.length-1;n>=0;n--){var o=r[n];if(0===o.path.length&&"replace"===o.op){t=o.value;break}}n>-1&&(r=r.slice(n+1));var i=g("Patches").$;return e(t)?i(t,r):this.produce(t,(function(t){return i(t,r)}))},n}(),it=new ot,ut=it.produce,at=it.produceWithPatches.bind(it),ct=it.setAutoFreeze.bind(it),ft=it.setUseProxies.bind(it),st=it.applyPatches.bind(it),lt=it.createDraft.bind(it),ht=it.finishDraft.bind(it);export{ot as Immer,st as applyPatches,U as castDraft,$ as castImmutable,lt as createDraft,I as current,W as enableAllPlugins,K as enableES5,N as enableMapSet,C as enablePatches,ht as finishDraft,v as freeze,L as immerable,e as isDraft,r as isDraftable,H as nothing,n as original,ut as produce,at as produceWithPatches,ct as setAutoFreeze,ft as setUseProxies};
-//# sourceMappingURL=immer.js.map
