--- public/web_modules/import-map.json
@@ -1,8 +1,0 @@
-{
-  "imports": {
-    "gsap": "./gsap.js",
-    "gsap/Draggable.js": "./gsap/Draggable.js",
-    "htm/preact/standalone.module.js": "./htm/preact/standalone.module.js",
-    "immer": "./immer.js",
-  }
-}
