+++ src/game/backend.js
@@ -1,4 +1,4 @@
-import {isDungeonCompleted} from '../game/utils-state.js'
+import {isDungeonCompleted} from './utils-state.js'
 
 const apiUrl = 'https://api.slaytheweb.cards/api/runs'
 // const apiUrl = 'http://localhost:3000/api/runs'
