+++ src/ui/animations.js
@@ -1,6 +1,5 @@
-import gsap from '../../public/web_modules/gsap.js'
-import {Draggable} from '../../public/web_modules/gsap/Draggable.js'
-// import {Flip} from '../web_modules/gsap/Flip.js'
+import gsap from 'gsap'
+import {Draggable} from 'gsap/Draggable.js'
 import Flip from 'https://slaytheweb-assets.netlify.app/gsap/Flip.js'
 
 // This file contains some resuable animations/effects.
