+++ src/ui/cards.js
@@ -1,5 +1,5 @@
 import {html, Component} from '../main.js'
-import {canPlay} from '../../public/game/conditions.js'
+import {canPlay} from '../game/conditions.js'
 
 export default class Cards extends Component {
 	// props = {gameState: {}, type ''}
