+++ src/ui/sounds.js
@@ -1,4 +1,3 @@
-// import * as Tone from '../web_modules/tone.js'
 import * as Tone from 'tone'
 
 // Create synths and connect it to the main output (your speakers).
