+++ src/ui/styles/app.css
@@ -10,7 +10,7 @@ html {
 	color: var(--text);
 	cursor: url(/images/cursors/default.png), auto;
 	min-height: 100%;
-	background: url('/images/scenes/dalle1.jpg');
+	background: url('/images/scenes/dalle5.jpg');
 	background-attachment: fixed;
 	/* background-position: center bottom; */
 	background-position: 50%;
