+++ tests/dungeon.js
@@ -1,10 +1,10 @@
 // @ts-ignore
 import test from 'ava'
-import {createTestDungeon} from '../public/content/dungeon-encounters.js'
-import actions from '../public/game/actions.js'
-import Dungeon from '../public/game/dungeon.js'
-import {MonsterRoom, Monster} from '../public/game/dungeon-rooms.js'
-import {getCurrRoom, isCurrRoomCompleted, isDungeonCompleted} from '../public/game/utils-state.js'
+import {createTestDungeon} from '../src/content/dungeon-encounters.js'
+import actions from '../src/game/actions.js'
+import Dungeon from '../src/game/dungeon.js'
+import {MonsterRoom, Monster} from '../src/game/dungeon-rooms.js'
+import {getCurrRoom, isCurrRoomCompleted, isDungeonCompleted} from '../src/game/utils-state.js'
 
 const a = actions
