+++ tests/game.js
@@ -1,7 +1,7 @@
 // @ts-ignore
 import test from 'ava'
-import createNewGame from '../public/game/new-game.js'
-import {createCard} from '../public/game/cards.js'
+import createNewGame from '../src/game/new-game.js'
+import {createCard} from '../src/game/cards.js'
 
 // We don't want to test too much here,
 // since tests/actions.js has most of it.
